% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_mats.R
\name{as.data.table.input_mats}
\alias{as.data.table.input_mats}
\title{Convert input matrices to \code{data.table}}
\usage{
\method{as.data.table}{input_mats}(x, ...)
}
\arguments{
\item{x}{An \code{\link{input_mats}} object.}

\item{...}{Currently unused.}
}
\value{
A \code{\link{data.table}}.
}
\description{
Convert an \code{\link{input_mats}} object to a single \code{\link{data.table}} that contains all
columns from the \code{X} matrices and all ID variables.
}
\seealso{
See \code{\link[=input_mats]{input_mats()}} for an example.
}
