% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.table.R
\name{as.data.table.tparams_transprobs}
\alias{as.data.table.tparams_transprobs}
\title{Coerce to \code{data.table}}
\usage{
\method{as.data.table}{tparams_transprobs}(x)
}
\arguments{
\item{x}{A \link{tparams_transprobs} object.}
}
\value{
A \code{data.table} with one row for each transition probability matrix.
}
\description{
Creates a \code{data.table} that combines the transition probability matrices
and ID columns from a \link{tparams_transprobs} object. This is often useful for
debugging.
}
\seealso{
\link{tparams_transprobs}
}
