% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap a statistical model}
\usage{
bootstrap(object, B, ...)
}
\arguments{
\item{object}{A statistical model.}

\item{B}{Number of bootstrap replications.}

\item{...}{Further arguments passed to or from other methods. Currently unused.}
}
\value{
Sampled values of the parameters.
}
\description{
\code{bootstrap} is a generic function for generating bootstrap replicates of the parameters
of a fitted statistical model.
}
