% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea.R
\docType{package}
\name{cea}
\alias{cea}
\alias{cea-package}
\title{cea: an R package for cost-effectiveness analysis}
\description{
It has three main goals:
}
\details{
\itemize{
\item Identify the most important data manipulation verbs and make them
  easy to use from R.
\item Provide blazing fast performance for in-memory data by writing key
  pieces in C++ (using Rcpp)
\item Use the same interface to work with data no matter where it's stored,
  whether in a data frame, a data table or database.
}

To learn more about dplyr, start with the vignettes:
\code{browseVignettes(package = "dplyr")}
}

