% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_mats.R
\name{check_input_data}
\alias{check_input_data}
\title{Check input data argument for \code{create_input_mats}}
\usage{
check_input_data(input_data)
}
\arguments{
\item{input_data}{An object of class "expanded_hesim_data" returned by the function
\code{\link[=expand.hesim_data]{expand.hesim_data()}}.}
}
\value{
If all tests passed, returns nothing; otherwise, throws an exception.
}
\description{
Check that input data argument for \code{create_input_mats} exists and that it is
of the correct type.
}
