% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctstm.R
\name{CohortCtstm}
\alias{CohortCtstm}
\title{Cohort-level continuous time state transition model}
\format{
An \link[R6:R6Class]{R6::R6Class} object.
}
\description{
Simulate outcomes from a cohort-level continuous time state transition
model (CTSTM). The class supports "clock-forward" (i.e., Markov), and mixtures of
clock-age and clock-forward multi-state models as described in
\code{\link{IndivCtstmTrans}}.
}
\seealso{
The \code{\link{IndivCtstmTrans}} documentation
describes the class for the transition model and the \code{\link{StateVals}} documentation
describes the class for the cost and utility models. An \code{\link{IndivCtstmTrans}}
object is typically created using \code{\link[=create_IndivCtstmTrans]{create_IndivCtstmTrans()}}.

This will provide very similar predictions to \code{IndivCtstm}, albeit
without the individual-level Monte Carlo variation. One semantic
difference is that point masses in \code{IndivCtstm} are always
clock-reset (essentially fixed times in state), whereas in
\code{CohortCtstm} the point masses are always clock-forward (Markov).

There is currently one relevant vignette.
\code{vignette("markov-inhomogeneous-cohort-continuous")} shows how the
individual-based time inhomogeneous Markov model in
\code{vignette("markov-inhomogeneous-indiv")} can be developed as a
cohort-based simulation.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{trans_model}}{The model for health state transitions. Must be an object
of class \code{\link{IndivCtstmTrans}}.}

\item{\code{utility_model}}{The model for health state utility. Must be an object of
class \code{\link{StateVals}}.}

\item{\code{cost_models}}{The models used to predict costs by health state.
Must be a list of objects of class \code{\link{StateVals}}, where each element of the
list represents a different cost category.}

\item{\code{stateprobs_}}{An object of class \code{\link{stateprobs}} simulated using \verb{$sim_stateprobs()}.}

\item{\code{qalys_}}{An object of class \code{\link{qalys}} simulated using \verb{$sim_qalys()}.}

\item{\code{costs_}}{An object of class \code{\link{costs}} simulated using \verb{$sim_costs()}.}

\item{\code{times_}}{A vector of times.}

\item{\code{zero_tol_}}{A double for being close to zero (used to avoid hazard(0.0)).}

\item{\code{abs_tol_}}{A double for the absolute tolerane in the ODE solver.}

\item{\code{rel_tol_}}{A double for the relative tolerane in the ODE solver.}

\item{\code{delta_time_}}{A double for the interval width used to
discretise time for ' semi-Markov models}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CohortCtstm-new}{\code{CohortCtstm$new()}}
\item \href{#method-CohortCtstm-sim}{\code{CohortCtstm$sim()}}
\item \href{#method-CohortCtstm-summarize}{\code{CohortCtstm$summarize()}}
\item \href{#method-CohortCtstm-clone}{\code{CohortCtstm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortCtstm-new"></a>}}
\if{latex}{\out{\hypertarget{method-CohortCtstm-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{CohortCtstm} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortCtstm$new(
  trans_model = NULL,
  utility_model = NULL,
  cost_models = NULL,
  zero_tol = 1e-100,
  abs_tol = 1e-06,
  rel_tol = 1e-06,
  delta_time = 0.1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trans_model}}{The \code{trans_model} field.}

\item{\code{utility_model}}{The \code{utility_model} field.}

\item{\code{cost_models}}{The \code{cost_models} field.}

\item{\code{zero_tol}}{The \code{zero_tol_} field.}

\item{\code{abs_tol}}{The \code{abs_tol_} field.}

\item{\code{rel_tol}}{The \code{rel_tol_} field.}

\item{\code{delta_time}}{The \code{delta_time_} field.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{CohortCtstm} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortCtstm-sim"></a>}}
\if{latex}{\out{\hypertarget{method-CohortCtstm-sim}{}}}
\subsection{Method \code{sim()}}{
Simulate quality-adjusted life-years (QALYs) as a function of
\code{utility_model} and costs as a function of \code{cost_models}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortCtstm$sim(
  t = c(0, 1),
  dr_qalys = 0.03,
  dr_costs = 0.03,
  type = c("predict", "random"),
  lys = TRUE,
  progress = NULL,
  by_patient = TRUE,
  debug = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{Double vector of times to evaluate the model (including the time origin).}

\item{\code{dr_qalys}}{Double for the discount rate for QALYs (not annualised).}

\item{\code{dr_costs}}{Double for the discount rate for costs (not annualised).}

\item{\code{type}}{\code{"predict"} for mean values or \code{"random"} for random samples
as in \verb{$sim()} in \code{\link{StateVals}}.}

\item{\code{lys}}{If \code{TRUE}, then life-years are simulated in addition to QALYs.}

\item{\code{progress}}{Integer for the number of simulations to report progress;
defaults to NULL for no reporting.}

\item{\code{by_patient}}{Logical for whether to report separately by patient; defaults to TRUE.}

\item{\code{debug}}{Logical for whether to print out debugging information; defaults to FALSE}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of \code{self} with simulated output of
class \link{qalys} stored in \code{qalys_} and class \link{costs} stored in \code{costs_}..
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortCtstm-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-CohortCtstm-summarize}{}}}
\subsection{Method \code{summarize()}}{
Summarize costs and QALYs so that cost-effectiveness analysis can be performed.
See \code{\link[=summarize_ce]{summarize_ce()}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortCtstm$summarize(by_grp = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{by_grp}}{If \code{TRUE}, then costs and QALYs are computed by subgroup. If
\code{FALSE}, then costs and QALYs are aggregated across all patients (and subgroups).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortCtstm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CohortCtstm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortCtstm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
