% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtstm.R
\name{create_CohortDtstm}
\alias{create_CohortDtstm}
\alias{create_CohortDtstm.model_def}
\title{Create \code{CohortDtstm} object}
\usage{
create_CohortDtstm(object, ...)

\method{create_CohortDtstm}{model_def}(object, input_data, ...)
}
\arguments{
\item{object}{An object of the appropriate class.}

\item{...}{Further arguments passed to \code{CohortDtstmTrans$new()} in
\link{CohortDtstmTrans}.}

\item{input_data}{An object of class \link[=expand.hesim_data]{expanded_hesim_data}.}
}
\description{
A generic function for creating an object of class \link{CohortDtstm}.
}
