% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtstm.R
\name{create_CohortDtstmTrans}
\alias{create_CohortDtstmTrans}
\alias{create_CohortDtstmTrans.multinom_list}
\title{Create \code{CohortDtstmTrans} object}
\usage{
create_CohortDtstmTrans(object, ...)

\method{create_CohortDtstmTrans}{multinom_list}(
  object,
  input_data,
  trans_mat,
  n = 1000,
  point_estimate = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of the appropriate class.}

\item{...}{Further arguments passed to \code{CohortDtstmTrans$new()} in
\code{CohortDtstmTrans}.}
}
\description{
A generic function for creating an object of class \code{CohortDtstmTrans}.
}
