% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctstm.R
\name{create_IndivCtstmTrans}
\alias{create_IndivCtstmTrans}
\alias{create_IndivCtstmTrans.flexsurvreg_list}
\alias{create_IndivCtstmTrans.flexsurvreg}
\title{Create \code{IndivCtstmTrans} object}
\usage{
create_IndivCtstmTrans(object, data, trans_mat, n = 1000,
  point_estimate = FALSE, ...)

\method{create_IndivCtstmTrans}{flexsurvreg_list}(object, data, trans_mat,
  n = 1000, point_estimate = FALSE, ...)

\method{create_IndivCtstmTrans}{flexsurvreg}(object, data, trans_mat,
  n = 1000, point_estimate = FALSE, ...)
}
\arguments{
\item{object}{A fitted statistical model.}

\item{data}{An object of class "expanded_hesim_data" returned by 
\code{\link{expand_hesim_data}}.}

\item{trans_mat}{The transition matrix describing the states and transitions in a 
multi-state model in the format from the \link[mstate]{mstate} package. See \link{IndivCtstmTrans}.}

\item{n}{Number of random observations of the parameters to draw.}

\item{point_estimate}{If \code{TRUE}, then the point estimates are returned and and no samples are drawn.}

\item{...}{Further arguments passed to \code{IndivCtstmTrans$new()} in \code{\link{IndivCtstmTrans}}.}
}
\value{
Returns an \code{\link{R6Class}} object of class \code{\link{IndivCtstmTrans}}.
}
\description{
A generic function for creating an object of class \code{\link{IndivCtstmTrans}}.
}
\seealso{
\code{\link{IndivCtstmTrans}}
}
