% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-data.R
\name{create_input_data}
\alias{create_input_data}
\title{Create input data}
\usage{
create_input_data(object, data, ...)
}
\arguments{
\item{object}{An object of the appropriate class. Currently supports
\code{\link{formula_list}}, \code{\link{lm}}, \code{\link{flexsurvreg}}, 
 \code{\link{flexsurvreg_list}}, and \code{\link{partsurvfit}}.}

\item{data}{An object of class "expanded_hesim_data" returned by the function
\code{\link{expand_hesim_data}}. Used to look for the input variables needed to create an input matrix
for use in a statistical models and the id variables for indexing rows in the input matrix.}

\item{...}{Further arguments passed to \code{\link{model.matrix}}.}
}
\value{
An object of class \code{\link{input_data}}.
}
\description{
\code{create_input_data} is a generic function for creating an object of class
\code{\link{input_data}}. Model matrices are constructed based on the 
variables specified in the model \code{object} and the data specified in \code{data}.
}
\examples{
library("flexsurv")

dt_strategies <- data.frame(strategy_id = c(1, 2))
dt_patients <- data.frame(patient_id = seq(1, 3), 
                          age = c(45, 47, 60),
                          female = c(1, 0, 0),
                          group = factor(c("Good", "Medium", "Poor")))
dt_states <- data.frame(state_id =  seq(1, 3),
                        state_name = factor(paste0("state", seq(1, 3))))
hesim_dat <- hesim_data(strategies = dt_strategies,
                        patients = dt_patients,
                        states = dt_states)

# Class "lm"
expanded_dat <- expand_hesim_data(hesim_dat, by = c("strategies", "patients", "states"))
fit_lm <- stats::lm(costs ~ female + state_name, psm4_exdata$costs$medical)
input_dat <- create_input_data(fit_lm, expanded_dat)
class(input_dat)

# Class "flexsurvreg"
expanded_dat <- expand_hesim_data(hesim_dat, by = c("strategies", "patients"))
fit_wei <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ 1, 
                                 data = ovarian, dist = "weibull")
input_dat <- create_input_data(fit_wei, expanded_dat)
class(input_dat)
}
