% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hesim_data.R
\name{create_time_intervals}
\alias{create_time_intervals}
\title{Create time intervals}
\usage{
create_time_intervals(time_start)
}
\arguments{
\item{time_start}{A vector of starting times for each interval}
}
\value{
A \code{data.table} in the same format as \code{time_intervals} as
described in \code{\link{id_attributes}}.
}
\description{
Create a table of time intervals given a vector of starting times for each
interval. This would typically be passed to \code{\link{id_attributes}}.
}
\examples{
create_time_intervals(c(0, 3, 5))
}
\seealso{
\code{\link{id_attributes}}
}
