% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctstm3_exdata}
\alias{ctstm3_exdata}
\title{Example data for a 3-state continuous time state transition model}
\format{A list containing the following elements:
\itemize{
\item{transitions}{ A data frame containining the times at which patient transitions between health states based
on the dataset \link[mstate]{prothr} from the \code{mstate} package.}
\item{costs}{ A list of data frames. The first data frame contains summary medical cost estimates and the
second data frame contains drug cost data.}
\item{utility}{ A data frame of summary utility estimates.}
}}
\usage{
ctstm3_exdata
}
\description{
A collection of example datasets containing health state transition, 
costs, and utility data for a 3-state continuous time state transition model.
}
\section{Transitions data}{

The data frame has the following columns:
\describe{
\item{strategy_id}{Treatment strategy identification number.}
\item{patient_id}{Patient identification number.}
\item{from}{Starting state.}
\item{to}{Receiving state.}
\item{trans}{Transition number.}
\item{Tstart}{Starting time.}
\item{Tstop}{Transition time.}
\item{years}{Elapsed years between \code{Tstart} and \code{Tstop}.}
\item{status}{Status variable; 1=transition, 0=censored.}
}
}

\section{Cost data}{

The cost list contains two data frames. The first data frame contains summary data on
medical costs by gender and health state, and contains the following columns:
\describe{
 \item{gender}{Female or male}
  \item{state_id}{The health state identification number.}
  \item{mean}{Mean costs.}
  \item{se}{Standard error of medical costs.}
 }
 
The second data frame contains data on the drug costs associated with each treatment strategy.
\describe{
 \item{strategy_id}{The treatment strategy identification number.}
 \item{costs}{Annualized drug costs.}
 }
}

\section{Utility data}{

The data frame has the following columns:
\describe{
\item{state_id}{The health state identification number.}
\item{mean}{Mean costs.}
\item{se}{Standard error.}
}
}

\keyword{datasets}
