% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_analysis.R
\name{custom_table}
\alias{custom_table}
\title{Custom CEA summary table}
\usage{
custom_table(x, arm, grp, custom_vars, FUN)
}
\arguments{
\item{x}{Matrix containing information on outcome variables for each simulation and arm.}

\item{arm}{Name of column denoting treatment arm.}

\item{grp}{Name of columne denoting subgroup}

\item{custom_vars}{Name of custom variables to calculate summary statistic for.}

\item{FUN}{summary statistic function.}
}
\value{
data.table
}
\description{
Custom summary table for CEA
}
