% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{design_matrix.flexsurvreg}
\alias{design_matrix.flexsurvreg}
\title{Obtain design matrix used to fit a flexsurvreg model}
\usage{
\method{design_matrix}{flexsurvreg}(x, indices = NULL, ...)
}
\arguments{
\item{x}{Output from \link{flexsurvreg} from the package \link{flexsurv}.}

\item{indices}{Indices from desgin matrices to extract. If NULL, each row 
from the design matrix using to fit the \link{flexsurvreg} object 
is returned.}

\item{...}{Further arguments passed to or from other methods. Currently unused.}
}
\value{
Returns a list of matrices. Each matrix corresponds to a parameter
of the probability distribution.
}
\description{
Obtain design matrix used to fit a survival model using
\code{flexsurv::flexsurvreg} or \code{flexsurv::flexsurvspline}.
}
\examples{
library("flexsurv")
fit <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ age, 
                   data = ovarian, dist = "weibull")
fit.X <- design_matrix(fit)
names(fit.X)
head(fit.X$scale)
head(fit.X$shape)
}
