% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{design_matrix.list}
\alias{design_matrix.list}
\title{Obtain design matrix used to fit a list of flexsurvreg models}
\usage{
\method{design_matrix}{list}(x, indices = NULL, ...)
}
\arguments{
\item{x}{Output from \link{flexsurvreg} from the package \link{flexsurv}.}

\item{indices}{Indices from desgin matrices to extract. If NULL, each row 
from the design matrix using to fit the \link{flexsurvreg} object 
is returned.}

\item{...}{Further arguments passed to or from other methods. Currently unused.}
}
\value{
Returns a list of lists matrices. Each list is equivalent to the
output produced by \link{design_matrix.flexsurvreg}.
}
\description{
Obtain design matrix used to fit a survival model using
\code{flexsurv::flexsurvreg} or \code{flexsurv::flexsurvspline}.
}
\examples{
library("flexsurv")
fit1 <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ 1, 
                   data = ovarian, dist = "exponential")
fit2 <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ 1, 
                   data = ovarian, dist = "weibull")                   
fit.X <- design_matrix(list(fit1, fit2))
names(fit.X)
head(fit.X[[1]])
}
