% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{design_matrix}
\alias{design_matrix}
\title{Obtain design matrix used to fit a model}
\usage{
design_matrix(x, ...)
}
\arguments{
\item{x}{A fitted statistical model.}

\item{...}{Additional arguments affecting extracted design matrix.}
}
\value{
The form of \code{design_matrix} depends on the class of its argument. See the 
documentation of the particular methods for details of what is produced.
}
\description{
\code{design_matrix} is a generic function for obtaining the design matrix 
used to fit a statistical model.
}
\seealso{
\link{design_matrix.flexsurvreg}, \link{design_matrix.list}
}
