% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hesim_data.R
\name{expand}
\alias{expand}
\title{Expand object}
\usage{
expand(object, by, times)
}
\description{
A generic function for "expanding" an object. Only used for
\code{hesim_data} objects with \code{\link[=expand.hesim_data]{expand.hesim_data()}}.
}
\seealso{
\code{\link[=expand.hesim_data]{expand.hesim_data()}}
}
\keyword{internal}
