% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-data.R
\name{expand_hesim_data}
\alias{expand_hesim_data}
\title{Expand hesim_data}
\usage{
expand_hesim_data(data, by = c("strategies", "patients"))
}
\arguments{
\item{data}{An object of class \code{\link{hesim_data}}.}

\item{by}{A character vector of the names of the data tables in \code{\link{hesim_data}} to expand by.}
}
\value{
This function is similar to \code{\link{expand.grid}}, but works for data frames or data tables. 
Specifically, it creates a \code{data.table} from all combinations of the supplied tables in \code{data}. 
The supplied tables are determined using the \code{by} argument. The resulting dataset is sorted by 
prioritizing id variables as follows: (i) \code{strategy_id}, (ii) \code{line}, (iii) \code{patient_id},
(iv) the health-related id variable (either \code{state_id} or \code{transition_id}).
}
\description{
Expand the data tables from an object of class \code{\link{hesim_data}} into a single \code{\link{data.table}} 
of class "expanded_hesim_data". See the description of the return value for details on how this is done.
}
\examples{
dt.strategies <- data.frame(strategy_id = c(1, 2))
dt.patients <- data.frame(patient_id = seq(1, 3), age = c(65, 50, 75),
                          gender = c("Female", "Female", "Male"))
dt.lines <- lines_dt(list(c(1, 2, 5), c(1, 2)))
dt.states <- data.frame(state_id =  seq(1, 3),
                        state_var = c(2, 1, 9))
data.tables <- hesim_data(strategies = dt.strategies,
                          patients = dt.patients,
                          states = dt.states,
                          lines = dt.lines)
expand_hesim_data(data.tables, by = c("strategies", "patients"))
}
