% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fits.R
\name{flexsurvreg_list}
\alias{flexsurvreg_list}
\title{List of \code{flexsurvreg} objects}
\usage{
flexsurvreg_list(...)
}
\arguments{
\item{...}{Objects of class \code{\link[flexsurv:flexsurvreg]{flexsurvreg}}, which can be named.}
}
\value{
An object of class \code{flexsurvreg_list}.
}
\description{
Combine \code{\link[flexsurv:flexsurvreg]{flexsurvreg}} objects into a list.
}
\examples{
library("flexsurv")
fit1 <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ 1, data = ovarian, dist = "weibull")
fit2 <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ 1, data = ovarian, dist = "exp")
fsreg_list <- flexsurvreg_list(wei = fit1, exp = fit2)
class(fsreg_list)
}
