% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-data.R
\name{form_input_data}
\alias{form_input_data}
\title{Form input data}
\usage{
form_input_data(object, data, ...)
}
\arguments{
\item{object}{An object of the appropriate class. Currently supports
\code{\link{formula_list}}, \code{\link{lm}}, \code{\link{flexsurvreg}}, 
 \code{\link{flexsurvreg_list}}, and \code{\link{partsurvfit}}.}

\item{data}{An object of class "expanded_hesim_data" returned by the function
\code{\link{expand_hesim_data}}. Used to look for the input variables needed to create an input matrix
for use in a statistical models and the id variables for indexing rows in the input matrix.}

\item{...}{Further arguments passed to \code{\link{model.matrix}}.}
}
\value{
An object of class \code{\link{input_data}}.
}
\description{
\code{form_input_data} is a generic function for forming data that can be
used as a input to a statistical model. Model matrices are formed based on the 
variables specified in the model \code{object} and the data specified in \code{data}.
}
\examples{
library("flexsurv")

dt.strategies <- data.frame(strategy_id = c(1, 2))
dt.patients <- data.frame(patient_id = seq(1, 3), 
                          age = c(45, 47, 60),
                          female = c(1, 0, 0),
                          group = factor(c("Good", "Medium", "Poor")))
dt.states <- data.frame(state_id =  seq(1, 3),
                        state_name = factor(paste0("state", seq(1, 3))))
hesim.dat <- hesim_data(strategies = dt.strategies,
                        patients = dt.patients,
                        states = dt.states)

# Class "lm"
expanded.dat <- expand_hesim_data(hesim.dat, by = c("strategies", "patients", "states"))
fit.lm <- stats::lm(costs ~ female + state_name, part_surv4_simdata$costs$medical)
input.dat <- form_input_data(fit.lm, expanded.dat)
class(input.dat)

# Class "flexsurvreg"
expanded.dat <- expand_hesim_data(hesim.dat, by = c("strategies", "patients"))
fit.wei <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ 1, 
                                 data = ovarian, dist = "weibull")
input.dat <- form_input_data(fit.wei, expanded.dat)
class(input.dat)
}
