% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{form_object_list}
\alias{form_object_list}
\title{Form a list from \code{...}}
\usage{
form_object_list(...)
}
\arguments{
\item{...}{Objects used to form a list.}
}
\value{
A list of objects from \code{...}.
}
\description{
Form a list of objects from \code{...}.
}
\keyword{internal}
