% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part-surv.R
\name{form_PartSurvCurves}
\alias{form_PartSurvCurves}
\title{Form \code{PartSurvCurves} object}
\usage{
form_PartSurvCurves(object, data, n = 1000, point_estimate = FALSE,
  bootstrap = TRUE)
}
\arguments{
\item{object}{An object of class \code{\link{partsurvfit}}.}

\item{data}{An object of class "expanded_hesim_data" returned by 
\code{\link{expand_hesim_data}}. Must be expanded by the data tables "strategies" and
"patients".}

\item{n}{Number of random observations of the parameters to draw.}

\item{point_estimate}{If \code{TRUE}, then the point estimates are returned and and no samples are drawn.}

\item{bootstrap}{If TRUE, then \code{n} bootstrap replications are drawn by refitting the survival
models in \code{object} on resamples of the sample data; if FALSE, then the parameters for each survival
model are independently draw from multivariate normal distributions.}

\item{...}{Further arguments passed to or from other methods. Currently unused.}
}
\value{
Returns an \code{\link{R6Class}} object of class \code{\link{PartSurvCurves}}.
}
\description{
\code{form_PartSurvCurves} is a generic function for forming an object of class
\code{\link{PartSurvCurves}} from a fitted statistical model.
}
