% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statevals.R
\name{form_StateVals}
\alias{form_StateVals}
\title{Form \code{StateVals} object}
\usage{
form_StateVals(object, data, n = 1000, point_estimate = FALSE)
}
\arguments{
\item{object}{A fitted statistical model object of the appropriate class. Supports
\code{\link{lm}}.}

\item{data}{An object of class "expanded_hesim_data" returned by 
\code{\link{expand_hesim_data}}. Must be expanded by the data tables "strategies",
"patients", and "states".}

\item{n}{Number of random observations of the parameters to draw.}

\item{point_estimate}{If \code{TRUE}, then the point estimates are returned and and no samples are drawn.}

\item{...}{Further arguments passed to or from other methods. Currently unused.}
}
\value{
Returns an \code{\link{R6Class}} object of class \code{\link{StateVals}}.
}
\description{
\code{form_StateVals} is a generic function for forming an object of class
 \code{\link{StateVals}} from a fitted statistical model.
}
