% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hesim_data.R
\name{get_id_size}
\alias{get_id_size}
\alias{get_id_size.input_mats}
\alias{get_id_size.expanded_hesim_data}
\alias{get_id_size.default}
\title{Get size of an ID attribute}
\usage{
get_id_size(x, ...)

\method{get_id_size}{input_mats}(x, id_name)

\method{get_id_size}{expanded_hesim_data}(x, id_name)

\method{get_id_size}{default}(x, id_name)
}
\arguments{
\item{x}{An object used to select a method.}

\item{...}{Further arguments passed to methods.}

\item{id_name}{The name of the ID variable to get the size for.}
}
\description{
A generic function for getting the size (i.e., number of unique values) of
an ID variable.
}
