% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea.R
\name{icea}
\alias{icea}
\alias{icea_pw}
\title{Individualized cost-effectiveness analysis}
\usage{
icea(x, ...)

icea_pw(x, ...)
}
\description{
These functions are deprecated, use \code{\link[=cea]{cea()}} and \code{\link[=cea_pw]{cea_pw()}} instead.
}
