% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hesim_data.R
\name{id_attributes}
\alias{id_attributes}
\title{Attributes for ID variables}
\usage{
id_attributes(strategy_id, n_strategies, patient_id, n_patients,
  state_id = NULL, n_states = NULL, transition_id = NULL,
  n_transitions = NULL, time_id = NULL, time_intervals = NULL,
  n_times = NULL, time_reset = NULL, time_fun = NULL)
}
\arguments{
\item{strategy_id}{A numeric vector of integers denoting the treatment strategy represented by each row
in \code{X}.}

\item{n_strategies}{A scalar denoting the number of unique treatment strategies.}

\item{patient_id}{A numeric vector of integers denoting the patient represented by each row
in \code{X}.}

\item{n_patients}{A scalar denoting the number of unique patients.}

\item{state_id}{A numeric vector of integers denoting the health state represented by each row
in \code{X}.}

\item{n_states}{A scalar denoting the number of unique health states.}

\item{transition_id}{A numeric vector denoting the 
health state transition represented by each row in \code{X}. This must only be specified when
estimating the health state transitions with a joint likelihood function. If independent
models are fit for each transition, then separate \code{X} matrices must be specified
for each transition. Note that this is not currently supported but
will be supported once \code{hesim} provides support for state transition modeling.}

\item{n_transitions}{A scalar denoting the number of unique transitions. 
Not supported by currently available models.}

\item{time_id}{A numeric vector of integers denoting a unique time interval.}

\item{time_intervals}{A \code{data.table} denotes unique time intervals. Must 
contain the columns \code{time_id}, \code{time_start}, and \code{time_stop}.
\code{time_start} is the starting time of an interval and \code{time_stop} is
the stopping time of an interval. Time intervals are closed on the left and
open on the right, and in the final interval, \code{time_stop} is equal to 
infinity.}

\item{n_times}{A scalar denoting the number of time intervals. Equal to the
number of rows in \code{time_intervals}.}

\item{time_reset}{If \code{TRUE}, then time intervals reset each time a patient enters a new health 
state. This is relevant if, for example, costs vary over time within health states. 
If \code{FALSE} then time intervals are based on time since the start of the simulation.}
}
\description{
Stores metadata related to the ID variables used to index input data and prediction objects.
}
\details{
The sorting order should be the same as specified in \code{\link{expand.hesim_data}}; that is,
observations must be sorted by: (i) \code{strategy_id}, (ii) \code{line}, 
(iii) \code{patient_id}, and (iv) the health-related ID variable (either \code{state_id} or
 \code{transition_id}).
 @export
}
