% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hesim_data.R
\name{id_getters}
\alias{id_getters}
\alias{get_n_strategies}
\alias{get_n_patients}
\alias{get_n_states}
\alias{get_n_transitions}
\alias{get_n_times}
\title{Generic getters}
\usage{
get_n_strategies(x, ...)

get_n_patients(x, ...)

get_n_states(x, ...)

get_n_transitions(x, ...)

get_n_times(x, ...)
}
\arguments{
\item{x}{An object used to select a method.}

\item{...}{Further arguments passed to methods.}
}
\description{
Generic functions for getting the size of ID variables from objects.
}
