% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea.R
\name{incr_change}
\alias{incr_change}
\title{Incremental changes}
\usage{
incr_change(x, control, sim, arm, outcomes)
}
\arguments{
\item{x}{Matrix containing information on outcome variables for each simulation and arm.}

\item{control}{Name of control arm.}

\item{sim}{Name of column denoting simulation number.}

\item{arm}{Name of column denoting treatment arm.}

\item{outcomes}{Name of columns to calculate incremental changes for.}
}
\value{
A data.table containing the differences in the values of each variable 
specified in outcomes between each treatment arm and the 
comparator. It is the same output generated
in \code{delta} from \code{pcea_pw}.
}
\description{
Calculate incremental change for outcome variables.
}
