% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-data.R
\name{input_data}
\alias{input_data}
\title{Input data for a statistical model}
\usage{
input_data(X, strategy_id, n_strategies, patient_id, n_patients, line = NULL,
  n_lines = NULL, state_id = NULL, n_states = NULL,
  transition_id = NULL, n_transitions = NULL, time_fun = NULL)
}
\arguments{
\item{X}{An input matrix or list of input matrices used for predicting
outcomes from a statistical model. The number of matrices depends on the statistical model
used.}

\item{strategy_id}{A numeric vector of integers denoting the treatment strategy represented by each row
in \code{X}.}

\item{n_strategies}{A scalar denoting the number of unique treatment strategies.}

\item{patient_id}{A numeric vector of integers denoting the patient represented by each row
in \code{X}.}

\item{n_patients}{A scalar denoting the number of unique patients.}

\item{line}{A numeric vector of integers denoting the treatment line represented by each row
in \code{X}. Not supported by currently available models.}

\item{n_lines}{A \code{\link{data.table}} denoting the number of treatment lines associated
with each treatment strategy. Should contain a column, "strategy_id", and a column,
"N". Not supported by currently available models.}

\item{state_id}{A numeric vector of integers denoting the health state represented by each row
in \code{X}.}

\item{n_states}{A scalar denoting the number of unique health states.}

\item{transition_id}{Only used for multi-state modeling. A numeric vector denoting the 
transition represented by each row in \code{X} (for jointly estimated models) or 
indexing distinct matrices in \code{X} (for separately estimated models). 
Not supported by currently available models.}

\item{n_transitions}{A scalar denoting the number of unique transitions. 
Not supported by currently available models.}

\item{time_fun}{A pointer to a C++ functor that can be used to update \code{X} as a function
of time in a simulation model. Not currently supported.}
}
\value{
An object of class "input_data", with elements equal to the specified 
function arguments.
}
\description{
Create an object of class "input_data", which contains data for use
as an input to a statistical model. The element \code{X} is the input matrix
used for prediction, and the id variables determine the dimensions for
 which predictions should be made (e.g,
by strategy and patient, by strategy, patient, and health state, ...). This object can 
be created more easily using the generic function \code{\link{form_input_data}}. See "details"
for information on soring the rows of the matrices in \code{X}.
}
\details{
The rows of the matrices in \code{X} must be sorted in a manner consistent with the id variables.
Sorting order should be the same as specified in \code{\link{expand_hesim_data}}; that is,
the rows of \code{X} must be sorted by prioritizing id variables as follows: \code{strategy_id},
\code{transition_id}, \code{patient_id}, \code{line}, and \code{state_id}.
}
\examples{
dt.strategies <- data.frame(strategy_id = c(1, 2))
dt.patients <- data.frame(patient_id = seq(1, 3), 
                          age = c(45, 47, 60),
                          female = c(1, 0, 0),
                          group = factor(c("Good", "Medium", "Poor")))
hesim.dat <- hesim_data(strategies = dt.strategies,
                             patients = dt.patients)

dat <- expand_hesim_data(hesim.dat, by = c("strategies", "patients"))$data
input.dat <- input_data(X = model.matrix(~ age, dat),
                       strategy_id = dat$strategy_id,
                       n_strategies = length(unique(dat$strategy_id)),
                       patient_id = dat$patient_id,
                       n_patients = length(unique(dat$patient_id)))
print(input.dat)
}
