% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lifetable}
\alias{lifetable}
\alias{lifetable_female}
\alias{lifetable_male}
\title{US Lifetables 2011}
\format{A data frame with 101 rows and 7 variables:
\describe{
  \item{age}{Age in years.}
  \item{qx}{Probability of dying between ages x and x + 1.}
  \item{lx}{Number surviving to age x.}
  \item{dx}{Number dying between ages x and x + 1.}
  \item{L}{Person-years lived between ages x and x + 1.}
  \item{Tx}{Total number of person-years lived above age x.}
  \item{ex}{Expectation of life at age x.}

}}
\source{
\url{http://www.cdc.gov/nchs/data/nvsr/nvsr64/nvsr64_11.pdf}
}
\usage{
lifetable_female

lifetable_male
}
\description{
Life tables by single-year of age from National Vital Statistics Reports
Volume 64, Number 11.
}
\keyword{datasets}
