% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-data.R
\name{lines_dt}
\alias{lines_dt}
\title{Data table of treatment lines}
\usage{
lines_dt(strategy_list, strategy_ids = NULL)
}
\arguments{
\item{strategy_list}{A list where each element is a treatment strategy 
consisting of a vector of treatments.}

\item{strategy_ids}{A numeric vector denoting the numeric id of each strategy
in \code{strategy_list}.}
}
\value{
Returns a \code{\link{data.table}} in tidy format with three columns
\describe{
\item{strategy_id}{Treatment strategy ids.}
\item{line}{Line of therapy.}
\item{treatment_id}{Treatment id for treatment used at a given line of therapy within a treatment strategy.}
}
}
\description{
Convert a list of treatment lines for multiple treatment strategies to a \code{\link{data.table}}
}
\examples{
strategies <- list(c(1, 2, 3),
                  c(1, 2))
lines_dt(strategies)
}
