% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fits.R
\name{lm_list}
\alias{lm_list}
\title{List of \code{lm} objects}
\usage{
lm_list(...)
}
\arguments{
\item{...}{Objects of class \code{\link[stats:lm]{lm}}, which can be named.}
}
\value{
Returns an object of class \code{lm_list}.
}
\description{
Combine \code{\link[stats:lm]{lm}} objects into a list
}
\examples{
 dat <- psm4_exdata$costs$medical
 lm_fits <- lm_list(fit1 = stats::lm(costs ~ 1, data = dat), 
                    fit2 = stats::lm(costs ~ female, data = dat))
 class(lm_fits)
}
\keyword{internal}
