% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpmatrix.R
\name{matrix_exp}
\alias{matrix_exp}
\title{Matrix exponential}
\usage{
matrix_exp(x, t = 1, ...)
}
\arguments{
\item{x}{A two-dimensional array like object where each row is a square
matrix ordered rowwise.}

\item{t}{An optional scaling factor for \code{x}.}

\item{...}{Arguments to pass to \link[msm:MatrixExp]{msm::MatrixExp}.}
}
\value{
Returns an array of exponentiated matrices.
}
\description{
This is a wrapper around \code{\link[msm:MatrixExp]{msm::MatrixExp()}} that computes the matrix
exponential of multiple square matrices.
}
\details{
This function is most useful when exponentiating transition intensity
matrices to produce transition probability matrices. To create transition
probability matrices for discrete time state transition models with annual
cycles, set \code{t=1}. An array of matrices is returned which can be used
to create the \code{value} element of a \link{tparams_transprobs} object. See
\code{\link[=qmatrix]{qmatrix()}} for an example.
}
\seealso{
\code{\link[=qmatrix]{qmatrix()}}
}
