% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{normboot}
\alias{normboot}
\alias{normboot.msm}
\title{Draw parameters of statistical model from multivariate normal distribution}
\usage{
normboot(object, B, ...)

\method{normboot}{msm}(object, B = 1000, ...)
}
\arguments{
\item{object}{A statistical model.}

\item{B}{Number of draws of the parameters.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
\code{normboot} is a generic function for drawing parameters from a fitted
statistical model from their (asymptotic) multivariate normal distribution.
}
\keyword{internal}
