% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter object}
\description{
Objects prefixed by "params_" are lists containing the parameters of a statistical model
used for simulation modeling. The parameters are used to simulate outcomes
as a function of covariates contained in input matrices (\link{input_mats}).
}
\seealso{
\code{\link{tparams}}
}
