% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params_joined_surv}
\alias{params_joined_surv}
\title{Parameters of joined survival models}
\usage{
params_joined_surv(..., times)
}
\arguments{
\item{...}{Objects of class \code{\link{params_surv}}, which can be named.}

\item{times}{A numeric vector of times at which to join models.}
}
\value{
An object of class "params_joined_surv", which is a list containing two elements:
\describe{
\item{models}{A list of \code{\link{params_surv}} objects from each statistical model
to be joined.}
\item{times}{Equivalent to the argument \code{times}.}
}
}
\description{
Create a list containing the parameters of survival models joined at specified time points. See
\code{\link{joined}} for more details.
}
\examples{
library("flexsurv")
fit.exp <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ 1, 
                                 data = ovarian, dist = "exp")
fit.wei <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ 1, 
                                 data = ovarian, dist = "weibull")
params.surv.exp <- form_params(fit.exp, n = 2)
params.surv.wei <- form_params(fit.wei, n = 2)
params.joined.surv <- params_joined_surv(exp = params.surv.exp,
                                         wei = params.surv.wei,
                                         times = 3)
print(params.joined.surv)
}
