% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params_joined_surv_list}
\alias{params_joined_surv_list}
\title{Parameters of joined lists of survival models}
\usage{
params_joined_surv_list(..., times)
}
\arguments{
\item{...}{Objects of class \code{\link{params_surv_list}}, which can be named.}

\item{times}{A list of sorted numeric vectors, with the length of each list element equal
to the number of sets of models.}
}
\value{
An object of class "params_joined_surv_list", which is a list containing two elements:
\describe{
\item{models}{A list of \code{\link{params_surv_list}}, each containing code{\link{params_surv}} 
objects to be joined.}
\item{times}{Equivalent to the argument \code{times}.}
}
}
\description{
Create a list containing the parameters of multiple sets of survival models, each joined
at specified time points.See \code{\link{joined}} for more details.
}
\examples{
library("flexsurv")
fit.exp <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ 1, 
                                 data = ovarian, dist = "exp")
fit.wei <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ 1, 
                                 data = ovarian, dist = "weibull")
fit.lnorm <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ 1, 
                                  data = ovarian, dist = "lognormal")

params.surv.exp <- form_params(fit.exp, n = 2)
params.surv.wei <- form_params(fit.wei, n = 2)
params.surv.lnorm <- form_params(fit.lnorm, n = 2)

params.surv.list1 <- params_surv_list(params.surv.exp, params.surv.wei)
params.surv.list2 <- params_surv_list(params.surv.exp, params.surv.lnorm)
params.joined.surv.list <- params_joined_surv_list(model1 = params.surv.list1,
                                                   model2 = params.surv.list2,
                                                   times = list(3, 5))
print(params.joined.surv.list)
}
