% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params_lm}
\alias{params_lm}
\title{Parameters of a linear model}
\usage{
params_lm(coefs, sigma = NULL)
}
\arguments{
\item{coefs}{Matrix of samples from the posterior distribution of the 
regression coefficients.}

\item{sigma}{A vector of samples of the standard error of the regression model.}
}
\value{
An object of class "params_lm", which is a list containing \code{coefs},
\code{sigma}, and \code{n_samples}. \code{n_samples} is equal to the number of rows
in \code{coefs}.
}
\description{
Create a list containing the parmeters of a fitted linear regression model.
}
\examples{
library("MASS")
n <- 2
params <- params_lm(coefs = MASS::mvrnorm(n, mu = c(.5,.6),
                                            Sigma = matrix(c(.05, .01, .01, .05), nrow = 2)),
                      sigma <- rgamma(n, shape = .5, rate = 4))
print(params)

}
