% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_evpi}
\alias{plot_evpi}
\title{Plot expected value of perfect information}
\usage{
plot_evpi(x, labels = NULL)
}
\arguments{
\item{x}{A \code{cea} object produced by \code{\link[=cea]{cea()}}.}

\item{labels}{A list of named vectors containing the levels and labels of
variables. The elements of each vector are
the levels of a variable and the names are the labels.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot the expected value of perfect information (EVPI) from the output of
\code{\link[=cea]{cea()}} using \code{\link{ggplot2}}. Intuitively, the EVPI provides an estimate of the
amount that a decision maker would be willing to pay to collect additional data
and completely eliminate uncertainty.
}
\details{
See the \code{\link[=cea]{cea()}} documentation for an example. If there are multiple subgroups,
then a faceted plot is produced with one plot for each subgroup.
}
