% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inlineCxxPlugin}
\alias{inlineCxxPlugin}
\title{Code to use the hesim package inline. Not directly called by the user.}
\usage{
inlineCxxPlugin(...)
}
\arguments{
\item{...}{arguments}
}
\description{
Code to use the hesim package inline. Not directly called by the user.
}
\examples{
  library(Rcpp)
  sourceCpp(code="
  // [[Rcpp::depends(hesim)]]
  // [[Rcpp::depends(RcppArmadillo)]]
  #include <hesim.h>
  // [[Rcpp::export]]
  double test_inline_gengamma(double mu, double sigma, double Q) {
    hesim::stats::gengamma gg(mu, sigma, Q);
    return gg.random();
  }")
  set.seed(12345)
  test_inline_gengamma(1.0, 1.0, 1.0)
}
\keyword{internal}
