% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_analysis.R
\name{psa}
\alias{psa}
\alias{psa_pw}
\title{Probabilistic sensitivity analysis}
\usage{
psa(x, k, sim = "sim", arm = "arm", grp = "grp", e = "e", c = "c",
  custom_vars = NULL, custom_fun = NULL)

psa_pw(x, k, control, sim = "sim", arm = "arm", grp = "grp", e = "e",
  c = "c", custom_vars = NULL, custom_fun = NULL)
}
\arguments{
\item{x}{Matrix containing information on mean costs and effectiveness for each simulation.
Should be in long form with unit of observation as the simulation and treatment arm.
Should have the following columns (
sim = simulation number,
arm = treatment arm,
c = summary of cost for each simulation and treatment arm
e = summary of clinicial effectiveness for each simulation and treatment arm)}

\item{k}{Vector of willingess to pay values}

\item{sim}{Name of column denoting simulation number. Default is "sim".}

\item{arm}{Name of column denoting treatment arm. Default is "arm".}

\item{grp}{Name of column denoting subgroup. Default is "grp".}

\item{e}{Name of column denoting clinical effectiveness. Default is "e".}

\item{c}{Name of column denoting costs. Default is "c".}

\item{custom_vars}{Character vector of variable names to use in creating a
custom summary table. Table will contain means and 95% credible intervals for
each variable. Can contain e and c.}

\item{custom_fun}{Function to apply to custom_vars to make custom table.}

\item{control}{Name of the control arm in x.}
}
\value{
list
}
\description{
Summarize output of probabilistic sensitivity analysis by subgroup.
}
