% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpmatrix.R
\name{qmatrix.msm}
\alias{qmatrix.msm}
\title{Transition intensity matrix from \code{msm} object}
\usage{
\method{qmatrix}{msm}(x, newdata, uncertainty = c("normal", "none"), n = 1000)
}
\arguments{
\item{x}{A \code{\link[msm:msm]{msm::msm}} object.}

\item{newdata}{A data frame to look for variables with which to predict. A
separate transition intensity matrix is predicted based on each row in
\code{newdata}.}

\item{uncertainty}{Method used to draw transition intensity matrices. If \verb{"none}",
then point estimates are used. If \code{"normal"}, then samples are drawn from the
multivariate normal distribution of the regression coefficients.}

\item{n}{Number of random observations of the parameters to draw.}
}
\value{
An array of transition intensity matrices with the third dimension
equal to the number of rows in \code{newdata}.
}
\description{
Draw transition intensity matrices for a probabilistic sensitivity analysis
from a fitted \code{msm} object.
}
\examples{
library("msm")
qinit <- rbind(
  c(-0.5, 0.25, 0, 0.25), 
  c(0.166, -0.498, 0.166, 0.166),
  c(0, 0.25, -0.5, 0.25), 
  c(0, 0, 0, 0)
)
fit <- msm(state ~ years, subject = PTNUM, data = cav,
           covariates =~ age,
           qmatrix = qinit, deathexact = 4)
qmatrix(fit,  n = 3)

}
\seealso{
\code{qmatrix.matrix()}
}
