% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.table.R, R/plot.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.data.table}
\alias{autoplot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table]{as.data.table}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}
}}

