% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{rsample.flexsurvspline}
\alias{rsample.flexsurvspline}
\title{Randomly sample parameters from a spline-based survival model}
\usage{
\method{rsample}{flexsurvspline}(x, n, ...)
}
\arguments{
\item{x}{Output from \link{flexsurvspline} from the package \link{flexsurv}.}

\item{n}{Number of random observations to draw.}

\item{...}{Further arguments passed to or from other methods. Currently unused.}
}
\value{
List where each element contains random samples of the coefficients
used to predict a given parameter of the survival distribution.
}
\description{
Draw random samples from a survival model fit using 
\code{flexsurv::flexsurvspline}.
}
\examples{
library("flexsurv")
fit <- flexsurv::flexsurvspline(Surv(recyrs, censrec) ~ group, data = bc, k=1, 
                     scale = "hazard")
fit.sample <- rsample(fit, n = 2)
head(fit.sample)
}
