% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{rsample.list}
\alias{rsample.list}
\title{Randomly sample parameters from a list of flexsurvreg models}
\usage{
\method{rsample}{list}(x, n, ...)
}
\arguments{
\item{x}{A list of models fit using \link{flexsurvreg} from the package
\link{flexsurv}.}

\item{n}{Number of random observations to draw.}

\item{...}{Further arguments passed to or from other methods. Currently unused.}
}
\value{
Returns an object of \link{class} "survlist_pars", which is a list 
of lists. Each list contains (i) random samples of the regression coefficients 
used to predict each of the model parameters and (ii) the
 probability distribution.
}
\description{
Draw random samples from a list of survival models fit using 
\code{flexsurv::flexsurvreg} or \code{flexsurv::flexsurvspline}.
}
\examples{
library("flexsurv")
fit1 <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ age, 
                    data = ovarian, dist = "exponential")
fit2 <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ 1, 
                   data = ovarian, dist = "gamma")
fits <- list(f1 = fit1, f2 = fit2)
fits.sample <- rsample(fits, n = 2)
head(fits.sample)
}
