% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{rsample}
\alias{rsample}
\title{Randomly sample parameters from a fitted model}
\usage{
rsample(x, n, ...)
}
\arguments{
\item{x}{A statistical model to randomly sample paraameters from.}

\item{n}{Number of random observations to draw.}

\item{...}{Additional arguments to pass to random sampling functions.}
}
\value{
The form of \code{rsample} depends on the class of its argument. See the 
documentation of the particular methods for details of what is produced.
}
\description{
\code{rsample} is a generic function for randomly sampling parameters from
a fitted statistical model. Parameters are sampled using the 
multivariate normal distribution.
}
\seealso{
\link{rsample.flexsurvreg}, \link{rsample.flexsurvspline}
}
