% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctstm.R
\name{run_CohortCtstmTestODE}
\alias{run_CohortCtstmTestODE}
\title{Test code for Markov cCTSTMs}
\usage{
run_CohortCtstmTestODE(
  start_state = 1,
  times = c(0, 1, 2, 10),
  discount_rate = 0.03
)
}
\description{
Test code for Markov cCTSTMs
}
\keyword{internal}
