% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_analyze.R
\name{sim_analyze}
\alias{sim_analyze}
\alias{sim_pv}
\alias{sim_los}
\alias{sim_transprob}
\title{Analyze simulated multi-state model}
\usage{
sim_pv(sim, tmat, r = 0.03, x = NULL, beta, poly.beta = NULL,
  poly.deg = NULL, knots = NULL, name = NULL, agevar = NULL)

sim_los(sim, timevar)

sim_transprob(sim, t)
}
\arguments{
\item{sim}{A simulated multi-state model from \code{\link{sim_msm}}.}

\item{r}{Discount rate for the model. Default is 0.03. Implies a discount factor of
\eqn{1/(1 + r)^t} where t is the cycle number.}

\item{x}{Data matrix}

\item{beta}{Matrix of regression coefficients corresponding to data matrix \code{x}.
One row for each state.}

\item{poly.beta}{Matrix of coefficeints for piecewise time polynomials. One row for each
state.}

\item{poly.deg}{Matrix indicating degree of polynomial on each interval. One row for each
state.}

\item{knots}{Matrix indicating location of knots for piecewise polynomials. One row for
each state.}
}
\value{
Vector of present value of quantity for each simulated
individual during time spent in state.
}
\description{
Analyze a simulated multi-state model
}
\keyword{internal}
