% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_msm.R
\name{sim_msm}
\alias{sim_msm}
\title{Simulate clock-reset multi-state model}
\usage{
sim_msm(loc_beta, loc_x, dist, tmat, par2, maxt, agevar = NULL)
}
\arguments{
\item{loc_beta}{Regression coefficients for location parameter. Array with 1st dimension (rows)
indexing random draw of coefficients, 2nd dimension (columns) indexing coefficients, and
3rd dimension (slice of cube) equal to number of unique transitions.}

\item{loc_x}{Data matrix for location parameter. Number of columns must equal length of
2nd dimension in loc_beta.}

\item{dist}{Character vector indicating the probability distributions used for each
transition. These include "weibull", "exponential", and "gompertz".}

\item{tmat}{Matrix indicating model transitions.}

\item{par2}{1st ancillary paramter (those other than the location parameter) in the model.}

\item{maxt}{Time to simulate model until.}

\item{agevar}{Name of age variable in loc_x.}
}
\value{
Dataframe with the following columns:
\item{id}{Identification number for each individual.}
\item{sim}{Simulation number (for each draw of parameters).}
\item{state}{Model state during each transition.}
\item{time}{Time state was reached.}
}
\description{
Simulate clock-reset multi-state model
}
\details{
The code is written in c++ to minmize simulation time.
}

