% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_param.R
\name{sim_param}
\alias{sim_param}
\title{Extract parameters for simulation from a multi-state model}
\usage{
sim_param(x)
}
\arguments{
\item{x}{A list of models fit with \code{flexsurvreg}}
}
\value{
List with the following parameters:
\item{loc_beta}{Regression coefficients for location parameter.}
\item{dist}{Distribution used for model fit.}
\item{par}{1st ancillary parameter in survvial model.}
}
\description{
Extract parameters for simulation from a multi-state model fit using \code{flexsurvreg}
from the \code{flexsurv} package
}

