% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-general.R
\name{sim_stateprobs}
\alias{sim_stateprobs}
\title{Simulated state probabilities}
\usage{
sim_stateprobs(x, ...)
}
\arguments{
\item{x}{An object of the appropriate class.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A \code{\link{stateprobs}} object.
}
\description{
A generic function to simulate state probabilities and create an object of
class \code{\link{stateprobs}}.
}
\seealso{
\code{\link{sim_stateprobs.survival}}
}
