% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statevals.R
\name{stateval_means}
\alias{stateval_means}
\title{Mean values for health states}
\usage{
stateval_means(values, strategy_id, patient_id)
}
\arguments{
\item{values}{Typically a matrix where each column denotes a health state and each
row denotes a random sample of the value assigned to that health state for the
probabilistic sensitivitiy analysis. If health state values vary by strategy, then
it can be an array of matrices where each matrix denotes the values associated with
a different treatment strategy.}

\item{strategy_id}{If a matrix, a vector denoting the strategy ID's to be modeled; if an array,
a vector denoting the strategy ID associated with each matrix in the array.}

\item{patient_id}{The patient ID's that will be modeled in the simulation. Note that state
values are assumed to be constant across patients when using objects of class "stateval_means".}
}
\value{
An object of class "stateval_means", which is a list containing \code{values},
\code{strategy_id}, and \code{patient_id}.
}
\description{
Estimated means that are used to simulate values assigned to health states with
 \code{\link{StateVals}}.
}
\examples{
# Cost estimates in 2 health states for a model with 2 treatment strategies and 3 patients
gamma_params <- mom_gamma(c(5000, 7000), c(1000, 1200))
n <- 3
vals <- matrix(rgamma(2 * n, 
                      shape = gamma_params$shape, 
                      scale = gamma_params$scale),
               nrow = n, ncol = 2, byrow = TRUE)
stval_ests <- stateval_means(values = vals,
                           strategy_id = c(1, 2),
                           patient_id = c(1, 2, 3))
print(stval_ests)
stateval_mod <- create_StateVals(stval_ests)
head(stateval_mod$sim(t = c(1, 2, 3), type = "predict"))

}
