% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_surv.R
\name{summary.params_surv}
\alias{summary.params_surv}
\title{Summary of a \code{params_surv} object}
\usage{
\method{summary}{params_surv}(object, prob = 0.95, ...)
}
\arguments{
\item{object}{An object of the appropriate class.}

\item{prob}{A numeric scalar in the interval \verb{(0,1)} giving the confidence
interval for coefficients. Default is 0.95 for a 95 percent interval, in which case
the lower and upper limits are computed using the 2.5th and 97.5th percentiles.}

\item{...}{Additional arguments affecting the summary. Currently unused.}
}
\description{
Summarize the coefficients of a parameter object by
computing the point estimate, lower confidence limit, and upper confidence
limit for each model term. The point estimate is the mean of the samples
of the coefficients and the lower and upper confidence limits are determined
by the \code{prob} argument. This is a convenient way to check whether a
a parameter object has been specified correctly and sampling distributions
of the coefficients are as expected.
}
\value{A \code{\link{data.table}} with the following columns:\describe{\item{model}{The name of the statistical model (e.g., a survival model for particular transition or endpoint).}
\item{parameter}{The name of the parameter of the survival distribution.}
\item{term}{The regression term.}
\item{estimate}{The estimated value of the regression term, computed as the mean fromits probability distribution.}
\item{lower}{The lower limit of the confidence interval for the estimate.}
\item{upper}{The upper limit of the confidence interval for the estimate.}}}
