% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tparams_transprobs.R
\name{summary.tparams_transprobs}
\alias{summary.tparams_transprobs}
\alias{print.tparams_transprobs}
\title{Summarize \code{tparams_transprobs} object}
\usage{
\method{summary}{tparams_transprobs}(object, probs = c(0.025, 0.975), unflatten = FALSE, ...)

\method{print}{tparams_transprobs}(x, ...)
}
\arguments{
\item{object}{A \code{\link{tparams_transprobs}} object.}

\item{probs}{A numeric vector of probabilities with values in \verb{[0,1]} used
to compute quantiles. By default, the 2.5th and 97.5th percentiles are
computed.}

\item{unflatten}{If \code{FALSE}, then each column containing a summary statistic
is a vector and the generated table contains one row
(for each set of ID variables) for each possible transition; if
\code{TRUE}, then each column stores a list of \code{matrix} objects containing
transition probability matrices formed by "unflattening" the one-dimensional
vectors. See "Value" below for additional details.}

\item{...}{For the \code{print()} method, arguments to pass to \code{summary.tparams_transprobs};
currently unused for the \code{summary()} method.}
}
\value{
\code{summary.tparams_transprobs()} uses the \code{\link[=summary.tpmatrix]{summary.tpmatrix()}}
method under the hood and returns the same output.
}
\description{
The \code{summary()} method summarizes a \code{\link{tparams_transprobs}} containing
predicted transition probabilities; summary statistics are computed for each
possible transition by the relevant ID variables. The \code{print()} method
summarizes the object using \code{summary.tparams_transprobs()} and prints it to the
console.
}
\seealso{
See \code{\link{tparams_transprobs}} for an example use of the summary and
print methods. Also see \code{\link[=summary.tpmatrix]{summary.tpmatrix()}} for more information on what is
returned.
}
