% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\docType{data}
\name{Survival}
\alias{Survival}
\title{Survival Simulation}
\format{\code{\link{R6Class}} object.}
\usage{
Survival
}
\description{
An R6 class for simulating health-economic models based on overall survival.
}
\section{Fields}{

\describe{
\item{\code{dist_name}}{Name of the probability distribution used to model survival.}

\item{\code{surv_coefs}}{An object of class "surv_pars" representing the posterior
distribution of coefficients from a survival model. For example, the output of
 \code{\link{rsample}}.}

\item{\code{surv_X}}{Matrix of covariates used to predict survival using the 
coefficients from \code{surv_coefs}.}
}}

\section{Members}{

 \describe{
  \item{\code{new()}}{Create a new object of class \code{Survival}}
  \item{\code{quantiles(q)}}{Generate the quantiles \code{q} of the survival 
  distribution. \code{q} is a numeric vector.} 
  \item{\code{surv(t)}}{Generate a survival curve give a numeric vector
  of times \code{t}.}
  \item{\code{cumhazard(t)}}{Generate cumulative hazards give a numeric vector
  of times \code{t}.}
  \item{\code{hazard(t)}}{Generate hazards give a numeric vector
  of times \code{t}.}
}
}

\keyword{datasets}
