% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{tparams}
\alias{tparams}
\title{Transformed parameter object}
\description{
Objects prefixed by "tparams_" are lists containing transformed parameters used
to simulate outcomes. The parameters have presumably already been transformed as a
function of input data and consequently do not need to be used alongside
input matrices. In other words, transformed parameters are parameters that have
already been predicted as a function of covariates.
}
\seealso{
\code{\link{params}}
}
