% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{tparams_transprobs}
\alias{tparams_transprobs}
\title{Transition probabilities}
\usage{
tparams_transprobs(value, ...)
}
\arguments{
\item{value}{Array of predicted transition probability matrices from the posterior
distribution of the underlying model.}

\item{...}{Arguments passed to \code{\link{id_attributes}}. Each matrix in 
\code{value} must be a prediction for a \code{sample}, \code{strategy_id},
 \code{patient_id}, and optionally \code{time_id} combination.}
}
\value{
An object of class "params_transprobs", which is a list containing 
\code{value} and the ID attributes passed to \code{\link{id_attributes}}.
}
\description{
Create a list containing predicted transition probabilities at discrete times
from a statistical model. Since the transition probabilities have presumably
already been predicted based on covariate values, no input data is required for
simulation.
}
