% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_def.R
\name{uv_rng}
\alias{uv_rng}
\title{Generate variates for univariate distributions}
\usage{
uv_rng(n, params, rng_fun, mom_fun = NULL, names = NULL)
}
\description{
Generate variates for univariate distributions
}
\keyword{internal}
