% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{absorbing}
\alias{absorbing}
\title{Absorbing states}
\usage{
absorbing(trans_mat)
}
\arguments{
\item{trans_mat}{A transition matrix in the format from the \link[mstate]{mstate} package. 
See \link{IndivCtstmTrans}.}
}
\description{
Returns a vector of absorbing states from a transition matrix.
}
\keyword{internal}
