% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\alias{bootstrap.partsurvfit}
\title{Bootstrap a statistical model}
\usage{
bootstrap(object, B, ...)

\method{bootstrap}{partsurvfit}(object, B, max_errors = 0, ...)
}
\arguments{
\item{object}{A statistical model.}

\item{B}{Number of bootstrap replications.}

\item{...}{Further arguments passed to or from other methods. Currently unused.}

\item{max_errors}{Maximum number of errors that are allowed when fitting statistical models
during the bootstrap procedure. This argument may be useful if, for instance, the model
fails to converge during some bootstrap replications. Default is 0.}
}
\value{
Sampled values of the parameters.
}
\description{
\code{bootstrap} is a generic function for generating bootstrap replicates of the parameters
of a fitted statistical model.
}
