% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-data.R, R/params.R, R/utils.R
\name{check.input_data}
\alias{check.input_data}
\alias{check.params_lm}
\alias{check.params_lm_list}
\alias{check.params_surv}
\alias{check.params_joined_surv}
\alias{check.params_surv_list}
\alias{check.params_joined_surv_list}
\alias{check}
\title{Input validation for class objects}
\usage{
\method{check}{input_data}(object)

\method{check}{params_lm}(object)

\method{check}{params_lm_list}(object)

\method{check}{params_surv}(object)

\method{check}{params_joined_surv}(object, inner_class)

\method{check}{params_surv_list}(object)

\method{check}{params_joined_surv_list}(object, inner_class)

check(object, ...)
}
\arguments{
\item{object}{object to check.}

\item{inner_class}{When checking a list of objects, the class of elements within
the inner most list.}

\item{...}{Further arguments passed to or from other methods. Currently unused.}
}
\value{
If validation is successful, returns the object in question; otherwise,
informs the user that an error has occurred.
}
\description{
\code{check} is a generic function for validating the inputs of class objects.
}
\keyword{internal}
