% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-data.R
\name{check_edata}
\alias{check_edata}
\title{Check data argument for \code{create_input_data}}
\usage{
check_edata(data)
}
\arguments{
\item{data}{An object of class "expanded_hesim_data" returned by the function
\code{\link{expand.hesim_data}}.}
}
\value{
If all tests passed, returns nothing; otherwise, throws an exception.
}
\description{
Check that data argument for \code{create_input_data} exists and that it is
of the correct type.
}
