% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtstm.R
\name{CohortDtstmTrans}
\alias{CohortDtstmTrans}
\title{Transitions for a cohort discrete time state transition model}
\format{An \link[R6:R6Class]{R6::R6Class} object.}
\description{
Simulate health state transitions in a cohort discrete time state transition model.
}
\seealso{
\code{\link[=create_CohortDtstmTrans]{create_CohortDtstmTrans()}}, \link{CohortDtstm}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{params}}{Parameters for simulating health state transitions. Currently
supports objects of class \link{tparams_transprobs}.}

\item{\code{input_mats}}{An object of class \link{input_mats}.}

\item{\code{start_stateprobs}}{A vector with length equal to the number of
health states containing the probability that the cohort is in each health
state at the start of the simulation. For example,
if there were three states and the cohort began the simulation in state 1,
then \code{start_stateprobs = c(1, 0, 0)}. If `NULL``, then a vector with the
first element equal to 1 and all remaining elements equal to 0.}

\item{\code{cycle_length}}{The length of a model cycle in terms of years.
The default is \code{1} meaning that model cycles are 1 year long.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CohortDtstmTrans$new()}}
\item \href{#method-sim_stateprobs}{\code{CohortDtstmTrans$sim_stateprobs()}}
\item \href{#method-clone}{\code{CohortDtstmTrans$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{CohortDtstmTrans} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortDtstmTrans$new(
  params,
  input_mats = NULL,
  start_stateprobs = NULL,
  cycle_length = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{The \code{params} field.}

\item{\code{input_mats}}{The \code{input_mats} field.}

\item{\code{start_stateprobs}}{The \code{start_stateprobs} field.}

\item{\code{cycle_length}}{The \code{cycle_length} field.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{CohortDtstmTrans} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sim_stateprobs"></a>}}
\subsection{Method \code{sim_stateprobs()}}{
Simulate probability of being in each health state during each model cycle.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortDtstmTrans$sim_stateprobs(n_cycles)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_cycles}}{The number of model cycles to simulate the model for.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \link{stateprobs}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortDtstmTrans$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
