% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtstm.R
\name{create_CohortDtstm}
\alias{create_CohortDtstm}
\alias{create_CohortDtstm.model_def}
\title{Create \code{CohortDtstm} object}
\usage{
create_CohortDtstm(object, ...)

\method{create_CohortDtstm}{model_def}(
  object,
  input_data,
  cost_args = NULL,
  utility_args = NULL,
  ...
)
}
\arguments{
\item{object}{An object of the appropriate class.}

\item{...}{Further arguments passed to \code{CohortDtstmTrans$new()} in
\code{\link{CohortDtstmTrans}}.}

\item{input_data}{An object of class
\code{\link[=expand.hesim_data]{expanded_hesim_data}}.}

\item{cost_args}{A list of further arguments passed to \code{StateVals$new()} in
\code{\link{StateVals}} when initiating cost models.}

\item{utility_args}{A list of further arguments passed to \code{StateVals$new()}
in \code{\link{StateVals}} when initiating the utility model.}
}
\description{
A generic function for creating an object of class \code{\link{CohortDtstm}}.
}
