% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtstm.R
\name{create_CohortDtstmTrans}
\alias{create_CohortDtstmTrans}
\title{Create \code{CohortDtstmTrans} object}
\usage{
create_CohortDtstmTrans(object, ...)
}
\arguments{
\item{object}{An object of the appropriate class.}

\item{...}{Further arguments passed to \code{CohortDtstmTrans$new()} in 
\code{CohortDtstmTrans}.}
}
\description{
A generic function for creating an object of class \code{CohortDtstmTrans}.
}
