% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtstm.R
\name{create_CohortDtstmTrans}
\alias{create_CohortDtstmTrans}
\alias{create_CohortDtstmTrans.multinom_list}
\alias{create_CohortDtstmTrans.msm}
\alias{create_CohortDtstmTrans.params_mlogit_list}
\title{Create \code{CohortDtstmTrans} object}
\usage{
create_CohortDtstmTrans(object, ...)

\method{create_CohortDtstmTrans}{multinom_list}(
  object,
  input_data,
  trans_mat,
  n = 1000,
  uncertainty = c("normal", "none"),
  ...
)

\method{create_CohortDtstmTrans}{msm}(
  object,
  input_data,
  cycle_length,
  n = 1000,
  uncertainty = c("normal", "none"),
  ...
)

\method{create_CohortDtstmTrans}{params_mlogit_list}(object, input_data, trans_mat, ...)
}
\arguments{
\item{object}{An object of the appropriate class.}

\item{...}{Further arguments passed to \code{CohortDtstmTrans$new()} in
\code{\link{CohortDtstmTrans}}.}

\item{input_data}{An object of class \code{expanded_hesim_data} returned by
\code{\link[=expand.hesim_data]{expand.hesim_data()}}}

\item{trans_mat}{A transition matrix describing the states and transitions
in a discrete-time multi-state model. See \code{\link{CohortDtstmTrans}}.}

\item{n}{Number of random observations to draw. Not used if \code{uncertainty = "none"}.}

\item{uncertainty}{Method determining how parameter uncertainty should be handled.
If \code{"normal"}, then parameters are randomly drawn from their multivariate normal
distribution. If \code{"none"}, then only point estimates are returned.}

\item{cycle_length}{The length of a model cycle in terms of years. The default
is 1 meaning that model cycles are 1 year long.}
}
\description{
A generic function for creating an object of class \code{\link{CohortDtstmTrans}}.
}
\seealso{
See \code{\link{CohortDtstmTrans}} for examples.
}
