% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtstm.R
\name{create_CohortDtstmTrans}
\alias{create_CohortDtstmTrans}
\alias{create_CohortDtstmTrans.multinom_list}
\title{Create \code{CohortDtstmTrans} object}
\usage{
create_CohortDtstmTrans(object, ...)

\method{create_CohortDtstmTrans}{multinom_list}(
  object,
  input_data,
  trans_mat,
  n = 1000,
  point_estimate = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of the appropriate class.}

\item{...}{Further arguments passed to \code{CohortDtstmTrans$new()} in
\code{CohortDtstmTrans}.}

\item{input_data}{An object of class \code{expanded_hesim_data} returned by
\code{\link[=expand.hesim_data]{expand.hesim_data()}}}

\item{trans_mat}{A transition matrix describing the states and transitions
in a discrete-time multi-state model. See \link{CohortDtstmTrans}.}

\item{n}{Number of random observations of the parameters to draw.}

\item{point_estimate}{If \code{TRUE}, then the point estimates are returned and and no samples are drawn.}
}
\description{
A generic function for creating an object of class \code{CohortDtstmTrans}.
}
