% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea.R
\name{custom_table}
\alias{custom_table}
\title{Custom CEA summary table}
\usage{
custom_table(x, strategy, grp, custom_vars, FUN = NULL)
}
\arguments{
\item{x}{Matrix containing information on outcome variables for each simulation and strategy.}

\item{strategy}{Name of column denoting treatment strategy.}

\item{grp}{Name of column denoting subgroup}

\item{custom_vars}{Name of custom variables to calculate summary statistic for.}

\item{FUN}{summary statistic function.}
}
\value{
A data.table of summary statistics for each variable specified in 
\code{custom_vars}. By default, returns the mean, 2.5\%, and 97.5\% quantile of
each variable. Different summary statistics can be calculated using FUN. 
This function is used in \code{icea} and \code{icea_pw} to create the
 \code{custom.table} output.
}
\description{
Custom table summarizing outcomes from probabilistic sensitivity analysis.
}
