% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-general.R
\name{disprog}
\alias{disprog}
\title{Disease progression object}
\description{
An object of class \code{disprog} returned from methods
\verb{$sim_disease()} in model classes. It contains simulated trajectories
through a multi-state model.
}
\section{Components}{

A \code{disprog} object inherits from \code{data.table} and contains
the following columns:

\describe{
\item{sample}{A random sample from the PSA.}
\item{strategy_id}{The treatment strategy ID.}
\item{patient_id}{The patient ID.}
\item{from}{The health state ID transitioned from.}
\item{to}{The health state ID transitioned to.}
\item{final}{An indicator equal to 1 if a patient is in their final health
state during the simulation and 0 otherwise.}
\item{time_start}{The time at the start of the interval.}
\item{time_stop}{The time at the end of the interval.}
}

The object also contains \code{size} and \code{absorbing} attributes.
The \code{size} attribute is a numeric vector with the elements \code{n_samples},
\code{n_strategies}, \code{n_patients}, \code{n_states} and \code{n_transitions} denoting the number of samples,
treatment strategies, patients, health states and transitions. The \code{absorbing} attribute
is a numeric vector containing the absorbing health states; i.e., the
health states that cannot be transitioned from. Operationally, an
absorbing state is a row in a transition matrix (as in the \code{trans_mat} field
of the \code{IndivCtstmTrans} class) with all \code{NA}s.
}

\seealso{
A disease progression object can be simulated with either the
\code{\link{IndivCtstm}} or \code{\link{IndivCtstmTrans}} classes.
}
