% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hesim_data.R
\name{expand}
\alias{expand}
\title{Expand object}
\usage{
expand(object, by)
}
\description{
A generic function for "expanding" an object. Only used for 
\code{\link{hesim_data}} objects.
}
\keyword{internal}
