% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fits.R
\name{formula_list}
\alias{formula_list}
\title{List of \code{formula} objects}
\usage{
formula_list(...)
}
\arguments{
\item{...}{Objects of class \code{\link{formula}}, which can be named.}
}
\value{
Returns an object of class "formula_list".
}
\description{
Create an object of class "formula_list". The object can be created from either
 multiple objects of class \code{\link{formula}} or from another "formula_list" object.
}
\examples{
# Create from "formula" objects
formula.list.wei <- formula_list(shape = formula(~ 1), scale = formula(~ x))
class(formula.list.wei)

# Create from "formula_list" objects
formula.list <- formula_list(exponential = formula_list(rate = formula(~1)),
                              weibull = formula.list.wei)

}
