% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fits.R
\name{formula_list}
\alias{formula_list}
\title{List of \code{formula} objects}
\usage{
formula_list(...)
}
\arguments{
\item{...}{Objects of class \link[stats:formula]{formula}, which can be named.}
}
\value{
An object of class \code{formula_list}.
}
\description{
Combine \link[stats:formula]{formula} or \link{formula_list} object into a
\link{formula_list} object.
}
\examples{
# Create from "formula" objects
flist_wei <- formula_list(shape = formula(~1), scale = formula(~x))
class(flist_wei)

# Create from "formula_list" objects
flist <- formula_list(
  exponential = formula_list(rate = formula(~1)),
  weibull = flist_wei
)

}
\keyword{internal}
