% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hesim.R
\docType{package}
\name{hesim}
\alias{hesim}
\title{\code{hesim}: Health Economic Simulation Modeling and Decision Analysis}
\description{
To learn more about \code{hesim} visit the \href{https://hesim-dev.github.io/hesim/}{website}.
}
