% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hesim.R
\docType{package}
\name{hesim}
\alias{hesim}
\alias{hesim-package}
\title{hesim: an R package for implement and analyzing health-economic simulation models}
\description{
hesim: an R package for implement and analyzing health-economic simulation models
}
