% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hesim.R
\docType{package}
\name{hesim}
\alias{hesim}
\title{hesim: an R package for high performance health-economic simulation modeling}
\description{
hesim: an R package for high performance health-economic simulation modeling
}
