% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea.R
\name{icea}
\alias{icea}
\alias{icea_pw}
\title{Individualized cost-effectiveness analysis}
\usage{
icea(x, k, sample = "sample", strategy = "strategy", grp = "grp",
  e = "e", c = "c")

icea_pw(x, k, comparator, sample = "sample", strategy = "strategy",
  grp = "grp", e = "e", c = "c")
}
\arguments{
\item{x}{A \code{data.frame} or \code{data.table} containing simulation output with
information on mean costs and effectiveness for each randomly sampled parameter set from
a PSA. Each row should denote a randomly sampled parameter set
and treatment strategy.}

\item{k}{Vector of willingness to pay values}

\item{sample}{Name of column denoting a randomly sampled parameter set. Default is "sample".}

\item{strategy}{Name of column denoting treatment strategy. Default is "strategy".}

\item{grp}{Name of column denoting subgroup. Default is "grp".}

\item{e}{Name of column denoting clinical effectiveness. Default is "e".}

\item{c}{Name of column denoting costs. Default is "c".}

\item{comparator}{Name of the comparator strategy in x.}
}
\value{
\code{icea} returns a list containing four data.tables:

\describe{
  \item{summary}{A data.table of the mean, 2.5\% quantile, and 97.5\% 
  quantile by strategy and group for clinical effectiveness and costs.}
  \item{mce}{The probability that each strategy is the most effective treatment
  for each group for the range of specified willingness to pay values.}
  \item{evpi}{The expected value of perfect information by group for the range
  of specified willingness to pay values.}
   \item{nmb}{The mean, 2.5\% quantile, and 97.5\% quantile of (monetary) net benefits
   for the range of specified willingness to pay values.}
}

\code{icea_pw} also returns a list containing four data.tables:
 \describe{
  \item{summary}{A data.table of the mean, 2.5\% quantile, and 97.5\% 
  quantile by strategy and group for clinical effectiveness and costs.}
  \item{delta}{Incremental effectiveness and incremental cost for each simulated
  parameter set by strategy and group. Can be used to plot a cost-effectiveness plane. }
  \item{ceac}{Values needed to plot a cost-effectiveness acceptability curve by
  group. In other words, the probability that each strategy is more cost-effective than
  the comparator for the specified willingness to pay values.}
   \item{inmb}{The mean, 2.5\% quantile, and 97.5\% quantile of (monetary) 
   incremental net benefits for the range of specified willingness to pay values.}
}
}
\description{
Conduct Bayesian cost-effectiveness analysis (e.g. summarize a probabilistic 
sensitivity analysis (PSA)) by subgroup.
\itemize{
 \item \code{icea()} compute the probability that
each treatment is most cost-effective, the expected value of perfect
information, and the net monetary benefit for each treatment.
\item \code{icea_pw()} compares interventions to a comparator. Computed
quantities include the incremental cost-effectiveness ratio, the 
incremental net monetary benefit, output for a cost-effectiveness plane,
and output for a cost-effectiveness acceptability curve.
}
}
\examples{
# simulation output
n_samples <- 100
sim <- data.frame(sample = rep(seq(n_samples), 4),
                  c = c(rlnorm(n_samples, 5, .1), rlnorm(n_samples, 5, .1),
                        rlnorm(n_samples, 11, .1), rlnorm(n_samples, 11, .1)),
                  e = c(rnorm(n_samples, 8, .2), rnorm(n_samples, 8.5, .1),
                        rnorm(n_samples, 11, .6), rnorm(n_samples, 11.5, .6)),
                  strategy = rep(paste0("Strategy ", seq(1, 2)),
                                 each = n_samples * 2),
                  grp = rep(rep(c("Group 1", "Group 2"),
                            each = n_samples), 2)
)

# icea
icea_dt <- icea(sim, k = seq(0, 200000, 500), sample = "sample", strategy = "strategy",
 grp = "grp", e = "e", c = "c")
names(icea_dt)
# the probability that each strategy is the most cost-effective 
# in each group with a willingness to pay of 20,000
library("data.table")
icea_dt$mce[k == 20000]

# icea_pw
icea_pw_dt <-  icea_pw(sim,  k = seq(0, 200000, 500), comparator = "Strategy 1",
                       sample = "sample", strategy = "strategy", e = "e", c = "c")
names(icea_pw_dt)
# cost-effectiveness acceptability curve
head(icea_pw_dt$ceac[k >= 20000])
}
