% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{icea}
\alias{icea}
\alias{icea_pw}
\title{Individualized cost-effectiveness analysis}
\usage{
icea(x, ...)

icea_pw(x, ...)
}
\arguments{
\item{x}{An object of simulation output characterizing the probability
distribution of clinical effectiveness and costs.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
These functions are deprecated, use \code{\link[=cea]{cea()}} and \code{\link[=cea_pw]{cea_pw()}} instead.
}
\keyword{internal}
