% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea.R
\name{icea}
\alias{icea}
\alias{icea_pw}
\alias{icea.default}
\alias{icea_pw.default}
\alias{icea.ce}
\alias{icea_pw.ce}
\title{Individualized cost-effectiveness analysis}
\usage{
icea(x, ...)

icea_pw(x, ...)

\method{icea}{default}(x, k = seq(0, 2e+05, 500), sample, strategy,
  grp = NULL, e, c, ...)

\method{icea_pw}{default}(x, k = seq(0, 2e+05, 500), comparator, sample,
  strategy, grp = NULL, e, c, ...)

\method{icea}{ce}(x, k = seq(0, 2e+05, 500), dr, ...)

\method{icea_pw}{ce}(x, k = seq(0, 2e+05, 500), comparator, dr, ...)
}
\arguments{
\item{x}{An object of simulation output characterizing the probability distribution
of clinical effectiveness and costs. If the default method is used, then \code{x}
must be a \code{data.frame} or \code{data.table} containing columns of
mean costs and clinical effectiveness where each row denotes a randomly sampled parameter set
and treatment strategy.}

\item{...}{Further arguments passed to or from other methods. Currently unused.}

\item{k}{Vector of willingness to pay values.}

\item{sample}{Character name of column from \code{x} denoting a randomly sampled parameter set.}

\item{strategy}{Character name of column from \code{x} denoting treatment strategy.}

\item{grp}{Character name of column from \code{x} denoting subgroup. If \code{NULL}, then
it is assumed that there is only one group.}

\item{e}{Character name of column from \code{x} denoting clinical effectiveness.}

\item{c}{Character name of column from \code{x} denoting costs.}

\item{comparator}{Name of the comparator strategy in x.}

\item{dr}{Discount rate.}
}
\value{
\code{icea} returns a list containing four \code{data.table}s:

\describe{
  \item{summary}{A \code{data.table} of the mean, 2.5\% quantile, and 97.5\% 
  quantile by strategy and group for clinical effectiveness and costs.}
  \item{mce}{The probability that each strategy is the most effective treatment
  for each group for the range of specified willingness to pay values. In addition,
  the column \code{best} denotes the optimal strategy (i.e., the strategy with the
  highest expected net monetary benefit), which can be used to plot the 
  cost-effectiveness acceptability frontier (CEAF).}
  \item{evpi}{The expected value of perfect information (EVPI) by group for the range
  of specified willingness to pay values. The EVPI is computed by subtracting the expected net
  monetary benefit given current information (i.e., the strategy with the highest
  expected net monetary benefit) from the expected net monetary benefit given
  perfect information.}
   \item{nmb}{The mean, 2.5\% quantile, and 97.5\% quantile of net monetary benefits
   for the range of specified willingness to pay values.}
}

\code{icea_pw} also returns a list containing four data.tables:
 \describe{
  \item{summary}{A data.table of the mean, 2.5\% quantile, and 97.5\% 
  quantile by strategy and group for clinical effectiveness and costs.}
  \item{delta}{Incremental effectiveness and incremental cost for each simulated
  parameter set by strategy and group. Can be used to plot a cost-effectiveness plane. }
  \item{ceac}{Values needed to plot a cost-effectiveness acceptability curve by
  group. The CEAC plots the probability that each strategy is more cost-effective than
  the comparator for the specified willingness to pay values.}
   \item{inmb}{The mean, 2.5\% quantile, and 97.5\% quantile of
   incremental net monetary benefits for the range of specified willingness to pay values.}
}
}
\description{
Conduct individualized cost-effectiveness analysis (ICEA) given output of an economic
model. That is, summarize a probabilistic sensitivity analysis (PSA) by subgroup.
\itemize{
 \item \code{icea()} computes the probability that
each treatment is most cost-effective, output for a cost-effectiveness acceptability frontier,
the expected value of perfect information, and the net monetary benefit for each treatment.
\item \code{icea_pw()} conducts pairwise ICEA by comparing strategies to a comparator. Computed
quantities include the incremental cost-effectiveness ratio, the 
incremental net monetary benefit, output for a cost-effectiveness plane,
and output for a cost-effectiveness acceptability curve.
}
}
\examples{
# simulation output
n_samples <- 100
sim <- data.frame(sample = rep(seq(n_samples), 4),
                  c = c(rlnorm(n_samples, 5, .1), rlnorm(n_samples, 5, .1),
                        rlnorm(n_samples, 11, .1), rlnorm(n_samples, 11, .1)),
                  e = c(rnorm(n_samples, 8, .2), rnorm(n_samples, 8.5, .1),
                        rnorm(n_samples, 11, .6), rnorm(n_samples, 11.5, .6)),
                  strategy = rep(paste0("Strategy ", seq(1, 2)),
                                 each = n_samples * 2),
                  grp = rep(rep(c("Group 1", "Group 2"),
                            each = n_samples), 2)
)

# icea
icea <- icea(sim, k = seq(0, 200000, 500), sample = "sample", strategy = "strategy",
             grp = "grp", e = "e", c = "c")
names(icea)
# The probability that each strategy is the most cost-effective 
# in each group with a willingness to pay of 20,000
library("data.table")
icea$mce[k == 20000]

# icea_pw
icea_pw <-  icea_pw(sim,  k = seq(0, 200000, 500), comparator = "Strategy 1",
                    sample = "sample", strategy = "strategy", grp = "grp",
                     e = "e", c = "c")
names(icea_pw)
# cost-effectiveness acceptability curve
head(icea_pw$ceac[k >= 20000])
icer_tbl(icea_pw)
}
