% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea.R
\name{icer}
\alias{icer}
\alias{format.icer}
\title{Incremental cost-effectiveness ratio}
\usage{
icer(x, prob = 0.95, k = 50000)

\method{format}{icer}(x, digit_qalys = 2, digit_costs = 0, pivot_from = "strategy")
}
\arguments{
\item{x}{An object of class \code{cea_pw} returned by \code{\link[=cea_pw]{cea_pw()}}.}

\item{prob}{A numeric scalar in the interval \verb{(0,1)} giving the confidence interval.
Default is 0.95 for a 95 percent interval.}

\item{k}{Willingness to pay quality-adjusted life-year.}

\item{pivot_from}{Character vector denoting a column or columns used to
"widen" the data. There will be one column for each value of the variables in
\code{pivot_from}. Default is to widen so there is a column for each treatment
strategy. Other options are the subgroup and outcome variable (or a combination of
them all). Set to \code{NULL} if you do not want to widen the table.}

\item{digits_qalys}{Number of digits to use to report QALYs.}

\item{digits_costs}{Number of digits to use to report costs.}
}
\value{
\code{icer()} returns a tidy \code{data.table} with the following columns:
\describe{
\item{strategy}{The treatment strategy.}
\item{grp}{The subgroup.}
\item{outcome}{The outcome metric.}
\item{mean}{The point estimate comptued as the average across the PSA samples.}
\item{lower}{The lower limit of the confidence interval.}
\item{upper}{The upper limit of the confidence interval.}
}

\code{format()} formats the table according to the arguments passed.
}
\description{
Generate a tidy table of incremental cost-effectiveness ratios (ICERs) given output from
\code{\link[=cea_pw]{cea_pw()}} with \code{icer()} and format for pretty printing with \code{format()}.
}
\seealso{
\code{\link[=cea_pw]{cea_pw()}}
}
