% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea.R
\name{incr_effect}
\alias{incr_effect}
\title{Incremental treatment effect}
\usage{
incr_effect(x, comparator, sim, strategy, grp, outcomes)
}
\arguments{
\item{x}{Matrix containing information on outcome variables for each simulation and strategy.}

\item{comparator}{Name of comparator strategy.}

\item{sim}{Name of column denoting simulation number.}

\item{strategy}{Name of column denoting treatment strategy.}

\item{grp}{Name of column denoting subgroup.}

\item{outcomes}{Name of columns to calculate incremental changes for.}
}
\value{
A data.table containing the differences in the values of each variable 
specified in outcomes between each treatment strategy and the 
comparator. It is the same output generated
in \code{delta} from \code{icea_pw}.
}
\description{
Calculate incremental effect of all treatment strategies on outcome variables from 
probabilistic sensitivity analysis relative to comparator.
}
