% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cea.R
\name{incr_effect}
\alias{incr_effect}
\title{Incremental treatment effect}
\usage{
incr_effect(x, control, sim, arm, outcomes)
}
\arguments{
\item{x}{Matrix containing information on outcome variables for each simulation and arm.}

\item{control}{Name of control arm (i.e. comparator).}

\item{sim}{Name of column denoting simulation number.}

\item{arm}{Name of column denoting treatment arm.}

\item{outcomes}{Name of columns to calculate incremental changes for.}
}
\value{
A data.table containing the differences in the values of each variable 
specified in outcomes between each treatment arm and the 
comparator. It is the same output generated
in \code{delta} from \code{pcea_pw}.
}
\description{
Calculate incremental effect of all treatment arms on outcome variables from 
probabilistic sensitivity analysis relative to comparator.
}
