% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctstm.R
\name{IndivCtstm}
\alias{IndivCtstm}
\title{Individual-level continuous time state transition model}
\format{An \link[R6:R6Class]{R6::R6Class} object.}
\description{
Simulate outcomes from an individual-level continuous time state transition
model (CTSTM) from a fitted multi-state model. The class supports "clock-reset"
(i.e., semi-Markov), "clock-forward" (i.e., Markov), and mixtures of
clock-reset and clock-forward models as described in
\link{IndivCtstmTrans}.
}
\seealso{
\code{\link[=create_IndivCtstmTrans]{create_IndivCtstmTrans()}}, \link{IndivCtstmTrans}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{trans_model}}{The model for health state transitions. Must be an object
of class \link{IndivCtstmTrans}.}

\item{\code{utility_model}}{The model for health state utility. Must be an object of
class \link{StateVals}.}

\item{\code{cost_models}}{The models used to predict costs by health state.
Must be a list of objects of class \link{StateVals}, where each element of the
list represents a different cost category.}

\item{\code{disprog_}}{A \code{data.table} simulated using \verb{$sim_disease()} containing the following columns:
\itemize{
\item \code{sample}: A random sample from the PSA.
\item \code{strategy_id}: The treatment strategy ID.
\item \code{patient_id}: The patient ID.
\item \code{from}: The health state ID transitioned from.
\item \code{to}: The health state ID transitioned to.
\item \code{final}: An indicator equal to 1 if a patient is in their final health
state during the simulation and 0 otherwise.
\item \code{time_start}: The time at the start of the interval.
\item \code{time_stop}: The time at the end of the interval.
}}

\item{\code{stateprobs_}}{An object of class \link{stateprobs} simulated using \verb{$sim_stateprobs()}.}

\item{\code{qalys_}}{An object of class \link{qalys} simulated using \verb{$sim_qalys()}.}

\item{\code{costs_}}{An object of class \link{costs} simulated using \verb{$sim_costs()}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{IndivCtstm$new()}}
\item \href{#method-sim_disease}{\code{IndivCtstm$sim_disease()}}
\item \href{#method-sim_stateprobs}{\code{IndivCtstm$sim_stateprobs()}}
\item \href{#method-sim_qalys}{\code{IndivCtstm$sim_qalys()}}
\item \href{#method-sim_costs}{\code{IndivCtstm$sim_costs()}}
\item \href{#method-summarize}{\code{IndivCtstm$summarize()}}
\item \href{#method-clone}{\code{IndivCtstm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{IndivCtstm} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndivCtstm$new(trans_model = NULL, utility_model = NULL, cost_models = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trans_model}}{The \code{trans_model} field.}

\item{\code{utility_model}}{The \code{utility_model} field.}

\item{\code{cost_models}}{The \code{cost_models} field.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{IndivCtstm} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sim_disease"></a>}}
\subsection{Method \code{sim_disease()}}{
Simulate disease progression (i.e., individual trajectories through a multi-state
model using an individual patient simulation).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndivCtstm$sim_disease(max_t = 100, max_age = 100, progress = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{max_t}}{A scalar or vector denoting the length of time to simulate the model.
If a vector, must be equal to the number of simulated patients.}

\item{\code{max_age}}{A scalar or vector denoting the maximum age to simulate each patient until.
If a vector, must be equal to the number of simulated patients.}

\item{\code{progress}}{An integer, specifying the PSA iteration (i.e., sample) that should be
printed every \code{progress} PSA iterations. For example, if \code{progress = 2},
then every second PSA iteration is printed. Default is \code{NULL},
in which case no output is printed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of self with simulated output stored in \code{disprog_}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sim_stateprobs"></a>}}
\subsection{Method \code{sim_stateprobs()}}{
Simulate health state probabilities as a function of time using the
simulation output stored in \code{disprog}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndivCtstm$sim_stateprobs(t)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{TA numeric vector of times.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of \code{self} with simulated output of class \link{stateprobs}
stored in \code{stateprobs_}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sim_qalys"></a>}}
\subsection{Method \code{sim_qalys()}}{
Simulate quality-adjusted life-years (QALYs) using \code{\link[=sim_qalys]{sim_qalys()}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndivCtstm$sim_qalys(
  dr = 0.03,
  type = c("predict", "random"),
  lys = TRUE,
  by_patient = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dr}}{Discount rate.}

\item{\code{type}}{\code{"predict"} for mean values or \code{"random"} for random samples
as in \verb{$sim()} in \link{StateVals}.}

\item{\code{lys}}{If \code{TRUE}, then life-years are simulated in addition to QALYs.}

\item{\code{by_patient}}{If \code{TRUE}, then QALYs are computed at the patient level.
If \code{FALSE}, then QALYs are averaged across patients by health state.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of \code{self} with simulated output stored
in \code{qalys_}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sim_costs"></a>}}
\subsection{Method \code{sim_costs()}}{
Simulate costs using \code{\link[=sim_costs]{sim_costs()}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndivCtstm$sim_costs(
  dr = 0.03,
  type = c("predict", "random"),
  by_patient = FALSE,
  max_t = Inf
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dr}}{Discount rate.}

\item{\code{type}}{\code{"predict"} for mean values or \code{"random"} for random samples
as in \verb{$sim()} in \link{StateVals}.}

\item{\code{by_patient}}{If \code{TRUE}, then QALYs are computed at the patient level.
If \code{FALSE}, then QALYs are averaged across patients by health state.}

\item{\code{max_t}}{Maximum time duration to compute costs once a patient has
entered a (new) health state. By default, equal to \code{Inf},
so that costs are computed over the entire duration that a patient is in
a given health state. If time varies by each cost category, then time
can also be passed as a numeric vector of length equal to the number of
cost categories (e.g., \code{c(1, 2, Inf, 3)} for a model with four cost categories).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of \code{self} with simulated output stored
in \code{costs_}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize"></a>}}
\subsection{Method \code{summarize()}}{
Summarize costs and QALYs so that cost-effectiveness analysis can be performed.
See \code{\link[=summarize_ce]{summarize_ce()}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndivCtstm$summarize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IndivCtstm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
