% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fits.R
\name{lm_list}
\alias{lm_list}
\title{List of \code{lm} objects}
\usage{
lm_list(...)
}
\arguments{
\item{...}{Objects of class \code{\link{lm}}, which can be named.}
}
\value{
Returns an object of class "lm_list".
}
\description{
Return an object of class "lm_list" multiple objects of class
\code{\link{lm}}.
}
\examples{
 dat <- part_surv4_simdata$costs$medical
 lm.list <- lm_list(fit1 = stats::lm(costs ~ 1, data = dat), 
                    fit2 = stats::lm(costs ~ female, data = dat))
 class(lm.list)
}
\keyword{internal}
