% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params_mlogit_list}
\alias{params_mlogit_list}
\title{Parameters of a list of multinomial logit models}
\usage{
params_mlogit_list(...)
}
\arguments{
\item{...}{Objects of class \code{\link{params_mlogit}}, which can be named.}
}
\value{
An object of class \code{params_mlogit_list}, which is a list containing
\code{\link{params_mlogit}} objects.
}
\description{
Create a list containing the parameters of multiple fitted multinomial logit models.
}
