% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params_surv}
\alias{params_surv}
\title{Parameters of a survival model}
\usage{
params_surv(coefs, dist, aux = NULL)
}
\arguments{
\item{coefs}{A list of length equal to the number of parameters in the 
survival distribution. Each element of the list is a matrix of samples from 
the posterior distribution of the regression coefficients used to predict
a given parameter.}

\item{dist}{Character vector denoting the parametric distribution. See "Details".}

\item{aux}{Auxillary arguments used with splines. See "Details".}
}
\value{
An object of class "params_surv", which is a list containing \code{coefs},
\code{dist}, and \code{n_samples}. \code{n_samples} is equal to the number of rows
in each element of \code{coefs}, which must be the same. The list may also contain \code{aux}, if
a spline model is fit.
}
\description{
Create a list containing the parameters of a single fitted parametric survival model.
}
\details{
The types of distributions that can be specified are: 
\itemize{
\item{\code{exponential} or \code{exp}}{ Exponential distribution}
\item{\code{weibull} or \code{weibull.quiet}}{ Weibull distribution}
\item{\code{gamma}}{ Gamma distribution}
\item{\code{lnorm}}{ Lognormal distribution}
\item{\code{gompertz}}{ Gompertz distribution}
\item{\code{llogis}}{ Log-logistic distribution}
\item{\code{gengamma}}{ Generalized gamma distribution}
\item{\code{survspline}}{ Survival splines}
}

Auxillary arguments for spline models should be specified as a list containing the elements:
\describe{
\item{\code{knots}}{ A numeric vector of knots.}
\item{\code{scale}}{ A character vector of length 1 denoting the survival outcome to be modeled
as a spline function. Options are "log_cumhazard", for the log cummulative hazard; 
"log_hazard" for the log hazard rate;
"log_odds" for the log cummulative odds;
and "normal" for the inverse normal distribution function.}
\item{\code{timescale}}{If "log" (the default), the survival outcome is modeled as a spline function
of log time; if "identity", it is modeled as a spline function of time.}
}
}
\examples{
library("flexsurv")
fit <- flexsurvreg(Surv(futime, fustat) ~ 1, data = ovarian, dist = "weibull")
params.surv <- params_surv(coefs = list(shape = fit$res.t["shape", "est", drop = FALSE],
                                        scale = fit$res.t["scale", "est", drop = FALSE]),
                           dist = fit$dlist$name)
print(params.surv)
}
