% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params_surv_list}
\alias{params_surv_list}
\title{Parameters of a list of survival models}
\usage{
params_surv_list(...)
}
\arguments{
\item{...}{Objects of class \code{\link{params_surv}}, which can be named.}
}
\value{
An object of class "params_surv_list", which is a list containing \code{params_surv}
objects.
}
\description{
Create a list containing the parameters of multiple fitted parametric survival models.
}
\examples{
library("flexsurv")
fit.wei <- flexsurvreg(Surv(futime, fustat) ~ 1, data = ovarian, dist = "weibull")
params.surv.wei <- form_params(fit.wei, n = 2)

fit.exp <- flexsurvreg(Surv(futime, fustat) ~ 1, data = ovarian, dist = "exp")
params.surv.exp <- form_params(fit.exp, n = 2)

params.surv.list <- params_surv_list(wei = params.surv.wei, exp = params.surv.exp)
print(params.surv.list)
}
