% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpmatrix.R
\name{qmatrix}
\alias{qmatrix}
\title{Transition intensity matrix}
\usage{
qmatrix(q, trans_mat)
}
\arguments{
\item{q}{A two-dimensional object that can be passed to \code{\link[=as.matrix]{as.matrix()}} containing
elements of the transition intensity matrix. A column represents a transition
from state $r$ to state $s$. Each row represents elements of a different
transition intensity matrix. See "Details" for more information.}

\item{trans_mat}{Just as in \link{IndivCtstmTrans}, a transition matrix
describing the states and transitions in a multi-state model.}
}
\value{
Returns a \code{qmatrix} object where each row is a flattened transition
intensity matrix with elements ordered rowwise.
}
\description{
\code{qpmatrix()} creates transition intensity matrices where elements represent
the instantaneous risk of moving between health states.
}
\details{
The object \code{q} must only contain non-zero and non-diagonal elements
of a transition intensity matrix. The diagonal elements are automatically computed
as the negative sum of the other rows.
}
\examples{
# 3 state irreversible model
tmat <- rbind(c(NA, 1, 2),
              c(NA, NA, 3),
              c(NA, NA, NA)) 
q12 <- c(.8, .7)
q13 <- c(.2, .3)
q23 <- c(1.1, 1.2)
q <- data.frame(q12, q13, q23)
qmatrix(q, trans_mat = tmat)

}
\seealso{
\code{\link[=tpmatrix]{tpmatrix()}}
}
