% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpmatrix.R
\name{qmatrix}
\alias{qmatrix}
\title{Transition intensity matrix}
\usage{
qmatrix(x, ...)
}
\arguments{
\item{x}{An \code{R} object.}

\item{...}{Further arguments passed to or from other methods. Currently unused.}
}
\description{
A generic function for creating transition intensity matrices where
elements represent the instantaneous risk of moving between health states.
}
