% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{rsample.flexsurvreg}
\alias{rsample.flexsurvreg}
\title{Randomly sample parameters from a flexsurvreg model}
\usage{
\method{rsample}{flexsurvreg}(x, n = NULL, point_estimate = FALSE, ...)
}
\arguments{
\item{x}{Output from \link{flexsurvreg} from the package \link{flexsurv}.}

\item{n}{Number of random observations to draw.}

\item{point_estimate}{If TRUE, then the point estimate is used for prediction
and no samples are drawn.}

\item{...}{Further arguments passed to or from other methods. Currently unused.}
}
\value{
Returns an object of \link{class} "survlist_pars". A list with
the following components is returned:
\describe{
\item{coefs}{List where each element is a matrix of sampled values of 
regression coefficients used to predict a parameter of the probability 
distribution.}
\item{dist}{Name of the probability distribution.}
}
}
\description{
Draw random samples from a survival model fit using 
\code{flexsurv::flexsurvreg} or \code{flexsurv::flexsurvspline}.
}
\examples{
library("flexsurv")
fit <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ 1, 
                   data = ovarian, dist = "weibull")
n <- 5
fit.sample <- rsample(fit, n = n)
head(fit.sample$coefs)
fit.sample$dist
}
