% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{rsample.flexsurvreg}
\alias{rsample.flexsurvreg}
\title{Randomly sample parameters from a flexsurvreg model}
\usage{
\method{rsample}{flexsurvreg}(x, n, ...)
}
\arguments{
\item{x}{Output from \link{flexsurvreg} from the package \link{flexsurv}.}

\item{n}{Number of random observations to draw.}

\item{...}{Further arguments passed to or from other methods. Currently unused.}
}
\value{
Returns an object of \link{class} "surv_pars", which is a list 
containing (i) random samples of the regression coefficients used to predict
each of the model parameters and (ii) the probability distribution.
}
\description{
Draw random samples from a survival model fit using 
\code{flexsurv::flexsurvreg} or \code{flexsurv::flexsurvspline}.
}
\examples{
library("flexsurv")
fit <- flexsurv::flexsurvreg(formula = Surv(futime, fustat) ~ 1, 
                   data = ovarian, dist = "weibull")
n <- 5
fit.sample <- rsample(fit, n = n)
head(fit.sample)
}
