% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_stateprobs.R
\name{sim_stateprobs.survival}
\alias{sim_stateprobs.survival}
\title{State probabilities from survival curves}
\usage{
\method{sim_stateprobs}{survival}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{survival}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A \code{stateprobs} object.
}
\description{
Simulate health state probabilities from a \code{\link{survival}} object using partitioned
survival analysis.
}
