% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{summarize_ce}
\alias{summarize_ce}
\title{Summarize costs and effectiveness}
\usage{
summarize_ce(costs, qalys)
}
\arguments{
\item{costs}{Simulated costs by category.}

\item{qalys}{Simulated QALYs.}
}
\value{
An object of class \code{\link{ce}}.
}
\description{
Summarize costs and quality-adjusted life-years (QALYs) given output simulated
from an economic model. The summary output is used to perform 
cost-effectiveness analysis with \code{\link{icea}} and \code{\link{icea_pw}}.
}
\details{
If mean costs and/or QALYs have already been computed 
(i.e., an average within a population), then there 
must be one observation for each discount rate (\code{dr}), 
PSA sample (\code{sample}), treatment strategy (\code{strategy_id}), 
and health state (\code{state_id}). Alternatively, there can be column
denoting a patient (\code{patient_id}), in which case outcomes will first be
averaged across patients.
}
\keyword{internal}
