% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summary.ce}
\alias{summary.ce}
\title{Summary method for cost-effectiveness object}
\usage{
\method{summary}{ce}(
  object,
  prob = 0.95,
  tidy = FALSE,
  digits_qalys = 2,
  digits_costs = 0,
  strategy_names = NULL,
  grp_names = NULL
)
}
\arguments{
\item{object}{A \code{\link{ce}} object.}

\item{prob}{A numeric scalar in the interval \verb{(0,1)} giving the confidence interval.
Default is 0.95 for a 95 percent interval.}

\item{tidy}{Whether to return a table in "tidy" format. See "Value".}

\item{digits_qalys}{Number of digits to use to report QALYs.}

\item{digits_costs}{Number of digits to use to report costs.}

\item{strategy_names}{Character vector of names for treatment strategies. Must
be the same length as the number of unique values of \code{strategy_id} in \code{object}.}

\item{grp_names}{Character vector of names for subgroups. Must
be the same length as the number of unique values of \code{grp_id} in \code{object}.}
}
\value{
A \code{data.table}. If \code{tidy = TRUE}, then the following columns are returned:

\describe{
\item{dr}{The discount rate.}
\item{strategy}{The treatment strategy.}
\item{grp}{The patient subgroup.}
\item{type}{Either \code{"QALYs"} or \code{"Costs"}.}
\item{category}{Category is always \code{"QALYs"} when \code{type == "QALYs"}; otherwise,
it is the cost category.}
\item{estimate}{The point estimate comptued as the average across the PSA samples.}
\item{lower}{The lower limit of the confidence interval.}
\item{upper}{The upper limit of the confidence interval.}
\item{value}{A formatted column that combines \code{estimate}, \code{lower}, and \code{upper} based
on \code{digit_qalys} and \code{digit_costs}.}
}
If \code{tidy = FALSE}, then the table above is reshaped into a wider format with one
column for each treatment strategy with values filled using the "value" column in the
tidy table. In this case the columns \code{estimate}, \code{lower}, and \code{upper} are removed and
there is one row for each discount rate and subgroup.
}
\description{
Summarize a \code{\link{ce}} object by producing confidence intervals for quality-adjusted
life-years (QALYs) and each cost category.
}
\details{
For an example, see \code{\link{IndivCtstm}}.
}
