% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R, R/params_surv.R, R/params_surv_list.R
\name{summary.params}
\alias{summary.params}
\alias{summary.params_surv}
\alias{summary.params_surv_list}
\title{Summarize parameter objects}
\usage{
\method{summary}{params_surv}(object, prob = 0.95, ...)

\method{summary}{params_surv_list}(object, prob = 0.95, ...)
}
\arguments{
\item{object}{An object of the appropriate class.}

\item{prob}{A numeric scalar in the interval \verb{(0,1)} giving the confidence
interval for coefficients. Default is 0.95 for a 95 percent interval, in which case
the lower and upper limits are computed using the 2.5th and 97.5th percentiles.}

\item{...}{Additional arguments affecting the summary. Currently unused.}
}
\value{
A \code{\link{data.table}} with the following columns:
\describe{
\item{parameter}{The name of the parameter of the survival distribution.}
\item{term}{The regression term.}
\item{estimate}{The estimated value of the regression term, computed as the mean
from its probability distribution.}
\item{lower}{The lower limit of the confidence interval for the estimate.}
\item{upper}{The upper limit of the confidence interval for the estimate.}
}
}
\description{
A collection of summary method for different parameter objects.
}
\seealso{
See \code{\link{params_surv}} for an example.
}
