% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R, R/params_mlogit.R,
%   R/params_mlogit_list.R, R/params_surv.R, R/params_surv_list.R
\name{summary.params}
\alias{summary.params}
\alias{summary.params_mlogit}
\alias{summary.params_mlogit_list}
\alias{summary.params_surv}
\alias{summary.params_surv_list}
\title{Summarize parameter objects}
\usage{
\method{summary}{params_mlogit}(object, prob = 0.95, ...)

\method{summary}{params_mlogit_list}(object, prob = 0.95, ...)

\method{summary}{params_surv}(object, prob = 0.95, ...)

\method{summary}{params_surv_list}(object, prob = 0.95, ...)
}
\arguments{
\item{object}{An object of the appropriate class.}

\item{prob}{A numeric scalar in the interval \verb{(0,1)} giving the confidence
interval for coefficients. Default is 0.95 for a 95 percent interval, in which case
the lower and upper limits are computed using the 2.5th and 97.5th percentiles.}

\item{...}{Additional arguments affecting the summary. Currently unused.}
}
\value{
A \code{\link{data.table}} containing at least the following columns:
\describe{
\item{term}{The regression term.}
\item{estimate}{The estimated value of the regression term, computed as the mean
from its probability distribution.}
\item{lower}{The lower limit of the confidence interval for the estimate.}
\item{upper}{The upper limit of the confidence interval for the estimate.}
}
In addition, the following columns may also be present:
\describe{
\item{parameter}{The name of the parameter of interest. This is relevant
for \code{\link{params_surv}} and \code{\link{params_surv_list}} objects since regression coefficients
are used to model the underlying parameters of the survival distribution (e.g.,
shape and scale for a Weibull model).}
\item{model}{The name of the statistical model. This is used for a
\code{\link{params_surv_list}} object, where each list element represents a separate model.
In a state transition model, each model is a unique health state transition and
in a partitioned survival model, there is a separate model for each curve.}
\item{to}{The health state that is being transitioned to. In \code{\link{params_mlogit}}
and \code{\link{params_mlogit_list}} objects, there are coefficients for each health
state that can be transitioned to.}
\item{from}{The health state that is being transitions from. This is used
for a \code{\link{params_mlogit_list}} objects where a different multinomial
logistic regression is used for each state that can be transitioned from.}
}
}
\description{
Summarize the coefficients of a parameter object by computing the point estimate,
lower confidence limit, and upper confidence limit for each model term. The point
estimate is the mean of the samples of the coefficients and the lower and
upper confidence limits are determined by the \code{prob} argument. This is a
convenient way to check whether a a parameter object has been specified
correctly and sampling distributions of the coefficients are as expected.
}
\seealso{
For examples, see the the underlying parameter object functions:
\code{\link[=params_surv]{params_surv()}}, \code{\link[=params_surv_list]{params_surv_list()}}, \code{\link[=params_mlogit]{params_mlogit()}}, and
\code{\link[=params_mlogit_list]{params_mlogit_list()}}.
}
