% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{survival}
\alias{survival}
\title{Survival object}
\description{
An object of class \code{survival} returned from \code{Psm$sim_survival()} or
\code{PsmCurves$survival()}.
}
\section{Components}{

A \code{survival} object inherits from \code{data.table} and contains
the following columns:

\describe{
\item{sample}{A random sample from the PSA.}
\item{strategy_id}{The treatment strategy ID.}
\item{patient_id}{The patient ID.}
\item{grp_id}{The subgroup ID.}
\item{curve}{One of the \code{N}-1 survival curves in an N-state partitioned
survival model. Each curve corresponds to unique endpoint.}
\item{t}{The time at which a survival probability is computed.}
\item{survival}{The probability of surviving to time \code{t}.}
}
}

\seealso{
\code{\link{Psm}}, \code{\link{PsmCurves}}
}
