% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\docType{data}
\name{Survival}
\alias{Survival}
\title{Survival Simulation}
\format{\code{\link{R6Class}} object.}
\usage{
Survival
}
\description{
An R6 class for simulating health-economic models based on overall survival.
}
\section{Fields}{

\describe{
\item{\code{dist_name}}{Name of the probability distribution used to model survival.}

\item{\code{surv_coefs}}{An object of class "surv_pars" representing the posterior
distribution of coefficients from a survival model. For example, the output of
 \code{\link{rsample}}.}

\item{\code{surv_X}}{Matrix of covariates used to predict survival using the 
coefficients from \code{surv_coefs}.}
}}

\keyword{datasets}
