% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtstm.R
\name{transprob_tbl}
\alias{transprob_tbl}
\title{Table to store transition probabilities}
\usage{
transprob_tbl(tbl, dist = c("custom", "dirichlet"), hesim_data = NULL)
}
\arguments{
\item{tbl}{A \code{data.frame} or \code{data.table} for storing parameter 
values. See "Details" for specifics.}

\item{dist}{Probability distribution used to sample parameters for a 
probabilistic sensitivity analysis (PSA).}

\item{hesim_data}{A \code{\link{hesim_data}} object. Required to specify 
patients if not included as columns in \code{tbl}, or, to match patients in 
\code{tbl} to groups. Not required if \code{tbl} includes one row for each
treatment strategy, patient, and transition combination. Patients are 
matched to groups by specifying both a \code{patient_id} and a 
\code{grp_id} column in the \code{patients} table;
if \code{hesim_data = NULL} but \code{grp_id} is included as a column
in \code{tbl}, then each group is assumed to be a unique patient.}
}
\value{
An object of class "transprob_tbl", which is a \code{data.table} of
parameter values with attributes for \code{dist} and optionally 
\code{strategy_id}, \code{patients}, and \code{transition_id}. \code{tbl} 
is in the same format as described in "Details". \code{patients} is a 
\code{data.table} with one column containing \code{patient_id} and 
optionally a second column containing \code{grp_id}.
}
\description{
Create a table for storing transition probabilities for a discrete time state
transition model stratified by treatment strategy, patient, transition, and
(optionally) time interval.
}
\details{
\code{tbl} is a \code{data.table} that must contain columns for
treatment strategies (\code{strategy_id}) and health state 
transitions (\code{transition_id}). The number of total transitions
must be equal to the square of number of health states. The transition
IDs are, by default, integers starting at 1 and ordered rowwise in 
transition matrices where each (i,j) pair denotes a transition from
state i to state j. For example, in a model with two health states 
transition 1 would be the transition from state 1 to itself, transition 2 would
be the transition from state 1 to state 2, transition 3 would be the
transition from state 2 to state 1, and transition 4 would be the transition 
from state 2 to itself. Alternatively, transition IDs can be specified in
\code{hesim_data} by including a \code{transitions} table. Each row is a 
treatment strategy and transition pair. 

Patient subgroups (\code{grp_id}) and time intervals (\code{time_start}) 
may optionally be included. If both are included, then each row is a treatment 
strategy, transition, patient subgroup, and time interval pair.

\code{tbl} must also contain columns summarizing the state values for each
row. Parameter values for a PSA can be drawn from Dirichlet distributions 
(separate for each treatment strategy and optionally patient subgroup 
and time interval) with \code{dist = 'dirichlet'}. Alternatively, 
\code{custom} can be used if transition probabilities for a PSA
 have been previously sampled from an arbitrary probability distribution.
 If \code{dist = "custom"}, then \code{tbl} must include a column 
 named \code{sample} (an integer vector denoting a unique random draw).
 The columns in \code{tbl} related to the values of the 
 transition probabilities depend on the value of \code{dist}. They are
 as follows:

\describe{
\item{dirichlet}{\code{alpha}}
\item{custom}{\code{value}}
}
}
\examples{
# hesim_data
strategies <- data.frame(strategy_id = c(1, 2))
patients <- data.frame(patient_id = seq(1, 3))
hesim_dat <- hesim_data(strategies = strategies,
                        patients = patients)

# Observed transitions 
# Treatment strategies 1 and 2
trans1 <- c(200, 300, 500,
            0, 400, 100,
            0, 0, 600) 
trans2 <- c(300, 300, 400,
            0, 550, 50,
            0, 0, 450)
n_trans <- length(trans1)
# Sample transition probabilities by treatment strategy
# from a dirichlet distributions
transprobs <- transprob_tbl(data.frame(strategy_id = rep(1:2, each = n_trans),
                                       transition_id = rep(1:n_trans, 2),
                                       alpha = c(trans1,trans2)),
                            dist = "dirichlet",
                            hesim_data = hesim_dat)
print(transprobs)
}
\seealso{
\code{\link{stateval_tbl}}
}
